% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive_ts.R
\name{archive_ts}
\alias{archive_ts}
\title{Archive a timeseries file}
\usage{
archive_ts(var_src, site_name, version = c("rds", "tsv"), ts_id, filename,
  verbose = TRUE)
}
\arguments{
\item{var_src}{the var_src for which to archive a file}

\item{site_name}{the site for which to archive a file}

\item{version}{the data storage format of the file to be archived}

\item{ts_id}{character. the ScienceBase ID[s] containing file[s] to archive}

\item{filename}{character. the filename[s] to archive}

\item{verbose}{logical. give status messages?}
}
\description{
Archiving means renaming a file to an archive name that reflects the file 
creation date. The file remains attached to the same timeseries item on 
ScienceBase.
}
\details{
For efficiency, e.g., when calling archive_ts() from another function that 
has already located the SB item, the ts_id and filename may be supplied with 
other arguments omitted. If ts_id and filename are omitted, the var_src, 
site, and version of the file to archive should be supplied instead, and this
function will locate the corresponding file[s] on ScienceBase.
}
