% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_calc_ts.R
\name{build_calc_ts_needs}
\alias{build_calc_ts_needs}
\title{Build the table calc_ts_needs.Rds and store it in extdata}
\usage{
build_calc_ts_needs(var, src)
}
\description{
Build the table calc_ts_needs.Rds and store it in extdata
}
\keyword{internal}

