% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_data.R
\name{build_sysdata}
\alias{build_sysdata}
\title{For use in package development only: build R/sysdata.rda}
\usage{
build_sysdata()
}
\description{
Moves/reformats the tables of variable and source metadata from inst/extdata
(where they can be manually edited in .tsv format) to R/sysdata.rda (the
storage location for data used internally by mda.streams)
}
\keyword{internal}

