% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_mfootdaily_calc3vK}
\alias{calc_mfootdaily_calc3vK}
\title{Internal - calculate the metabolic footprint for each day}
\usage{
calc_mfootdaily_calc3vK(utctime, K600, veloc, wtr)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC corresponding to the date vector}

\item{K600}{the daily estimates of K600}

\item{veloc}{the instantaneous velocities}

\item{date}{the Dates of K600}

\item{wtrdf}{the ts dataframe for instantaneous water temperatures (with 
DateTime column)}
}
\description{
From Chapra & DiToro 1991, between eq 1 & eq 2: "sufficiently long distance"
= >3U/ka, where U = stream velocity (m d^-1), ka = reaeration rate (d^-1)
}
\keyword{internal}
