% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_depth_calcDischHarvey}
\alias{calc_ts_depth_calcDischHarvey}
\title{Internal - calculate depth_calcDisch from discharge and depth-vs-discharge
coefficients from Jud Harvey}
\usage{
calc_ts_depth_calcDischHarvey(utctime, disch, c, f)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC/GMT}

\item{disch}{the discharge in ft^3 s^-1}

\item{c}{the multiplier in d = c * Q^f}

\item{f}{the exponent in d = c * Q^f}
}
\description{
Internal - calculate depth_calcDisch from discharge and depth-vs-discharge
coefficients from Jud Harvey
}
\keyword{internal}

