% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_doamp_calcDAmp}
\alias{calc_ts_doamp_calcDAmp}
\title{Internal - calculate the daily amplitude in DO}
\usage{
calc_ts_doamp_calcDAmp(sitetime, longitude, dopsat)
}
\arguments{
\item{sitetime}{the local time, e.g. from sitetime_calcLon}

\item{doobs}{the observed DO concentration}
}
\description{
Internal - calculate the daily amplitude in DO
}
\keyword{internal}

