% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_doinit_calcDStart}
\alias{calc_ts_doinit_calcDStart}
\title{Internal - calculate doinit_calcDStart from any data}
\usage{
calc_ts_doinit_calcDStart(sitetime, doobs, longitude, day_start = 4,
  day_end = 28)
}
\arguments{
\item{sitetime}{the DateTimes of the local noons of interest, in UTC}

\item{doobs}{vector of dissolved oxygen values, same length as sitetime}

\item{longitude}{the site longitude in degrees E}
}
\description{
Internal - calculate doinit_calcDStart from any data
}
\keyword{internal}

