% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_dopsat_calcObsSat}
\alias{calc_ts_dopsat_calcObsSat}
\title{Internal - calculate dopsat_calcObsSat from any data}
\usage{
calc_ts_dopsat_calcObsSat(utctime, doobs, dosat)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC/GMT}

\item{doobs}{the observed DO concentration}

\item{dosat}{the DO concentration at saturation}
}
\description{
Internal - calculate dopsat_calcObsSat from any data
}
\keyword{internal}

