% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_dosat_calcGG}
\alias{calc_ts_dosat_calcGG}
\title{Internal - calculate depth_calcDisch from any data}
\usage{
calc_ts_dosat_calcGG(utctime, wtr, baro)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC/GMT}

\item{wtr}{the water temperature}

\item{baro}{the barometric pressure}
}
\description{
Internal - calculate depth_calcDisch from any data
}
\keyword{internal}

