% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_par_calcLat}
\alias{calc_ts_par_calcLat}
\title{Internal - calculate par_calcLat from any data}
\usage{
calc_ts_par_calcLat(utctime, suntime, latitude)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC}

\item{suntime}{the apparent solar time at the site}

\item{latitude}{the site latitude in degrees N}
}
\description{
Internal - calculate par_calcLat from any data
}
\keyword{internal}
