% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_par_calcSw}
\alias{calc_ts_par_calcSw}
\title{Internal - calculate par from SW data}
\usage{
calc_ts_par_calcSw(utctime, sw)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC}

\item{sw}{shortwave radiation in W m^-2}
}
\description{
Internal - calculate par from SW data
}
\keyword{internal}

