% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_simCopy}
\alias{calc_ts_simCopy}
\title{Internal - calculate any variable with simNew from any data}
\usage{
calc_ts_simCopy(var, from_src, from_site, filter_fun)
}
\arguments{
\item{var}{the variable name}

\item{from_src}{the src for var to copy from}

\item{from_site}{the site to copy from}

\item{filter_fun}{NULL, or a function to apply to the copied data before 
staging it to a new file. filter_fun needs to accept and return a unitted
data.frame of the standard ts form for mda.streams}
}
\description{
Internal - calculate any variable with simNew from any data
}
\keyword{internal}

