% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_simNew}
\alias{calc_ts_simNew}
\title{Internal - calculate any variable with simNew from any data}
\usage{
calc_ts_simNew(var, utctime, value)
}
\arguments{
\item{var}{the variable name}

\item{utctime}{the DateTime with tz of UTC/GMT}

\item{value}{unitted vector of ts values}
}
\description{
Internal - calculate any variable with simNew from any data
}
\keyword{internal}

