% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_sitedate_calcLon}
\alias{calc_ts_sitedate_calcLon}
\title{Internal - calculate sitedate_calcLon from any data}
\usage{
calc_ts_sitedate_calcLon(sitetime, longitude)
}
\arguments{
\item{sitetime}{the DateTimes of the local noons of interest, in UTC/GMT}

\item{longitude}{the site longitude in degrees E}
}
\description{
Internal - calculate sitedate_calcLon from any data
}
\keyword{internal}

