% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_sitetime_calcLon}
\alias{calc_ts_sitetime_calcLon}
\title{Internal - calculate sitetime_calcLon from any data}
\usage{
calc_ts_sitetime_calcLon(utctime, longitude)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC/GMT}

\item{longitude}{the site longitude in degrees E}
}
\description{
Internal - calculate sitetime_calcLon from any data
}
\keyword{internal}

