% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_sitetimedaily_calcLon}
\alias{calc_ts_sitetimedaily_calcLon}
\title{Internal - calculate sitetimedaily_calcLon from any data}
\usage{
calc_ts_sitetimedaily_calcLon(sitetime, longitude)
}
\arguments{
\item{sitetime}{the DateTimes of the local noons of interest, in UTC/GMT}

\item{longitude}{the site longitude in degrees E}
}
\description{
Internal - calculate sitetimedaily_calcLon from any data
}
\keyword{internal}

