% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_veloc_calcDischHarvey}
\alias{calc_ts_veloc_calcDischHarvey}
\title{Internal - calculate velocity from discharge and velocity-vs-discharge
coefficients from Jud Harvey}
\usage{
calc_ts_veloc_calcDischHarvey(utctime, disch, k, m)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC/GMT}

\item{disch}{the discharge in ft^3 s^-1}

\item{k}{the multiplier in U = k * Q^m}

\item{m}{the exponent in U = k * Q^m}
}
\description{
Internal - calculate velocity from discharge and velocity-vs-discharge
coefficients from Jud Harvey
}
\keyword{internal}

