% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_velocdaily_calcDMean}
\alias{calc_ts_velocdaily_calcDMean}
\title{Internal - calculate daily mean velocity from instantaneous}
\usage{
calc_ts_velocdaily_calcDMean(sitetime, longitude, veloc)
}
\arguments{
\item{sitetime}{the local time, e.g. from sitetime_calcLon}
}
\description{
Daily means are computed for the 24 hour periods from midnight to midnight in
sitetime. Their DateTime values are noon in sitetime, converted back to UTC.
These centers approximately correspond to the daily centers for the
metabolism estimation data.
}
\keyword{internal}

