% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_with_input_check}
\alias{calc_ts_with_input_check}
\title{Helper to stage_calc_ts for simXxx-style functions}
\usage{
calc_ts_with_input_check(inputs, calc_fun)
}
\arguments{
\item{inputs}{a list of named inputs to check}

\item{calc_fun}{the name (character) of the function against which to check 
the inputs}
}
\description{
Checks to see that the inputs are just the right ones to supply to the
function named by calc_fun
}
