% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choose_data_source.R
\name{choose_data_source}
\alias{choose_data_source}
\title{Select the specific data source to use for a given variable and site}
\usage{
choose_data_source(var, site)
}
\arguments{
\item{var}{character. The variable name, e.g. "doobs", "disch", or "wtr"}
}
\value{
a two-column, one-row data.frame with variable-specific column names
  and cell values indicating the source type (e.g., "local", "proxy", "model") and
  specific source (e.g., "nwis_01474000", "bird")
}
\description{
Select the specific data source to use for a given variable and site
}
\examples{
choose_data_source("doobs", "nwis_03067510")
}

