% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choose_sites.R
\name{choose_sites}
\alias{choose_sites}
\title{Identify sites with sufficient data to run a metabolism model}
\usage{
choose_sites(required_vars = c("disch", "doobs", "wtr"))
}
\arguments{
\item{required_vars}{character vector of variable names that are all required
for a site to qualify}
}
\description{
We will need to be more flexible later about which variables are available,
since the file names will be variants on the core variable names (e.g.,
wtr.local, wtr.model, etc.)
}
\examples{
choose_sites()
}

