% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bundle_packages.R
\name{cluster_bundle_packages}
\alias{cluster_bundle_packages}
\title{Build a zip file of bundled packages}
\usage{
cluster_bundle_packages(packages = "mda.streams",
  github = "USGS-R/mda.streams@develop",
  bundlefile = "./cluster_packages.zip", pkgdir = "temp/pkgs",
  repos = union(getOption("repos"), "http://owi.usgs.gov/R"))
}
\arguments{
\item{packages}{character. the package or packages whose dependencies (and 
the packages themselves) should be obtained. at least for now, all packages
and dependencies should have an up-to-date list of dependencies on CRAN or 
GRAN. It's OK if the GitHub code is ahead of CRAN/GRAN as long as the
dependencies list hasn't changed}

\item{github}{character vector of GitHub references (as in 
\code{link[github]{install_github}}) for packages to acquire from GitHub 
rather than CRAN/GRAN}

\item{bundlefile}{file.path to save the bundled packages}

\item{pkgdir}{directory path where the downloaded and written packages should
be saved}

\item{repos}{character vector of repositories with info on the package 
dependencies}
}
\description{
To use this file, unpack it. It contains a top-level folder called packages. 
To install a package from this bundle, run 
\code{install.packages('mypackagename', repos='file:packages', 
type='source')}
}

