% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tables.R
\name{combine_dplyr}
\alias{combine_dplyr}
\title{A function to combine unitted data.frames/tbl_dfs}
\usage{
combine_dplyr(method, by, ...)
}
\arguments{
\item{method}{character. dplyr join function to use.}

\item{by}{columns to join on, to be passed to the dplyr join function and
arrange_}
}
\value{
a joined, sorted, unitted tbl_df
}
\description{
A function to combine unitted data.frames/tbl_dfs
}
\keyword{internal}

