% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_tables.R
\name{combine_tables}
\alias{combine_tables}
\title{Combine several tables into a single table}
\usage{
combine_tables(..., by, fun = combine_dplyr("full_join", by = by),
  allow_constants = FALSE)
}
\arguments{
\item{...}{tables (data.frames or tbl_dfs) to combine. These may be unitted 
if fun is prepared for unitted tables. Tables are joined left to 
right. e.g., if dots are A, B, C then the result is 
fun(fun(A,B),C)}

\item{by}{character. The columns[s] on which to join the tables. Good choices
are 'DateTime' for timeseries data or 'site_name' for site metadata.}

\item{fun}{the combining function to apply to pairs of tables in the iterative 
join.}

\item{allow_constants}{logical. if TRUE, tables with \code{NA} in their \code{by} column}
}
\value{
a joined function
}
\description{
Combine several tables into a single table
}
\keyword{internal}
