% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ts.R
\name{condense_by_ply}
\alias{condense_by_ply}
\title{Condense function called by mm_model_by_ply}
\usage{
condense_by_ply(data_ply, data_daily_ply, day_start, day_end, ply_date,
  ply_validity, timestep_days, ...)
}
\arguments{
\item{data_ply}{as in mm_model_by_ply_prototype}

\item{data_daily_ply}{as in mm_model_by_ply_prototype}

\item{day_start}{as in mm_model_by_ply_prototype}

\item{day_end}{as in mm_model_by_ply_prototype}

\item{ply_date}{as in mm_model_by_ply_prototype}

\item{ply_validity}{as in mm_model_by_ply_prototype}

\item{timestep_days}{as in mm_model_by_ply_prototype}

\item{...}{as in mm_model_by_ply_prototype}
}
\description{
Following mm_model_by_ply_prototype
}
\keyword{internal}
