% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_data.R
\name{config_preds_to_data_column}
\alias{config_preds_to_data_column}
\title{Get predictions of metabolism or DO from a model}
\usage{
config_preds_to_data_column(var, site, src, type)
}
\arguments{
\item{var}{the variable name to retrieve}

\item{src}{the model_name or model file name}

\item{type}{character in c("SB","file") indicating whether src is a model on 
sciencebase or a local file name}
}
\description{
Most of this code is the same whether it's a pred (the name of a model on 
sciencebase) or a pred_file (the name of a local file name)
}
\details{
The trick is that model predictions only store the input time, with was 
either local.time (< streamMetabolizer 0.8.0) or solar.time (>=0.8.0) rather 
than UTC time. So this function draws on our knowledge of the site-specific 
mapping between UTC time and solar.time or date (as stored in sitetime 
and sitedate, respectively) to come up with the UTC DateTime stamps that 
correspond to these predictions. We could potentially just use
convert_solartime_to_UTC, but I'm nervous about tiny imprecisions that would
make these not match up to other DateTime stamps.
}
\keyword{internal}

