% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_data.R
\name{config_to_data_column}
\alias{config_to_data_column}
\title{Read a single column worth of data}
\usage{
config_to_data_column(var, type, site, src, optional = FALSE)
}
\arguments{
\item{var}{character. A single variable name as in 
\code{unique(get_var_src_codes(out="var"))}.}

\item{type}{character in \code{c("ts","meta","file","const")}.}

\item{site}{character. The site code or codes for which to choose data.}

\item{src}{character indicating a src to be interpreted in the context of 
\code{type}}
}
\value{
a data.frame, else NA
}
\description{
For use within config_to_data
}
\keyword{internal}
