% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_to_metab_clusterfun.R
\name{config_to_metab_clusterfun}
\alias{config_to_metab_clusterfun}
\title{Create a function to pass to cluster nodes}
\usage{
config_to_metab_clusterfun(config, return_value = c("metab_model",
  "file_name", "nothing"), repeat_times = NA, post_metab = TRUE,
  stage_folder = NULL, sb_user, sb_password, verbose = TRUE)
}
\arguments{
\item{config}{a data.frame containing the configuration information, or the 
name of a config file on the computer cluster node.}

\item{return_value}{character. What should be returned when the function 
completes? "nothing" may be appropriate for many-model runs or large model 
objects.}

\item{repeat_times}{either NA for no repetition, or an integer number of 
times to repeat the fit for each config row (using config_to_metab_repeat)}

\item{post_metab}{logical. Should the metab model be staged and posted?}

\item{stage_folder}{character. The folder name on the computer cluster node 
where the metab_model should be staged before posting (only applies if 
post_metab=TRUE). If NULL, the local tempdir() will be used.}

\item{sb_user}{the ScienceBase username, only used if post_metab=TRUE}

\item{sb_password}{the ScienceBase password, only used if post_metab=TRUE}

\item{verbose}{logical. Should status messages be given?}
}
\description{
Returns a self-contained function (closure) with all the necessary 
information to run \code{\link{config_to_metab}} or 
\code{\link{config_to_metab_repeat}} on computer cluster nodes via 
\code{\link[parallel]{clusterApply}} or 
\code{\link[parallel]{clusterApplyLB}}
}
\examples{
\dontrun{
# might need to run this before calling clusterApplyLB, but might work all in one line
clusterApplyLB(c1, 1:nrow(config), config_to_metab_clusterfun(config, rows))
}
}

