% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config_to_metab_repeat.R
\name{config_to_metab_repeat}
\alias{config_to_metab_repeat}
\title{Acquire data and fit a model repeatedly}
\usage{
config_to_metab_repeat(config, row, times = 5, verbose = FALSE)
}
\arguments{
\item{config}{a config table}

\item{row}{a single integer specifying hte row of config to use}

\item{times}{integer giving the number of times to fit the one config row}

\item{verbose}{logical. give status messages?}
}
\value{
A metab_model object with an augmented \code{info} slot (access with
  \code{get_info}) that includes, most importantly, a list element named
  \code{fit_reps}: \code{fit_reps} is a data.frame containing information on
  each repetition of the config_to_metab procedure, one row per repetition.
  Data columns include the metabolism estimates, data prep time, model
  fitting time, doobs simulation model, and an abbreviated version of the
  metab estimation config specifications .
}
\description{
This function really only makes sense in the context of models that are
fitted from simulated doobs data, which in turn must be specified as preds
from a metab_sim model.
}

