% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delete_item.R
\name{delete_item}
\alias{delete_item}
\title{Delete a ScienceBase item}
\usage{
delete_item(item_ids, item_names, delete_files = FALSE,
  delete_children = FALSE, delete_item = TRUE, verbose = TRUE)
}
\arguments{
\item{item_ids}{character. SB id[s]}

\item{item_names}{character. The name[s] by which to refer to the item[s] in
messages}

\item{delete_files}{logical. Should files be sought and deleted from within
this item? If FALSE, files will be ignored.}

\item{delete_children}{logical. Should child items be sought and recursively
deleted from within this item? If FALSE, children will be ignored.}

\item{delete_item}{logical. Should the item itself be deleted after files
and/or children are deleted?}

\item{verbose}{logical. Should status messages be given?}
}
\description{
Deletes one or more generic items, including their child items and files as
applicable & specified.
}
\keyword{internal}

