% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_meta.R
\name{delete_meta}
\alias{delete_meta}
\title{Delete a metadata item-file combination}
\usage{
delete_meta(types, files_only = FALSE, verbose = TRUE)
}
\description{
Delete a metadata item-file combination
}

