% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delete_metab_run.R
\name{delete_metab_run}
\alias{delete_metab_run}
\title{Delete a metab_run item from SB}
\usage{
delete_metab_run(title, files_only = FALSE, verbose = TRUE)
}
\arguments{
\item{title}{the metab_run title}

\item{files_only}{logical. Only delete the files, leaving an empty run item?}

\item{verbose}{logical. Give status messages?}
}
\description{
Delete the run item, and/or its files, from SB
}

