% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_sb_file.R
\name{delete_sb_file}
\alias{delete_sb_file}
\title{Delete a single file from a ScienceBase item}
\usage{
delete_sb_file(sb_id, files, verbose = TRUE)
}
\arguments{
\item{sb_id}{character sciencebase id}

\item{files}{character filename[s] to delete}

\item{verbose}{logical. Should status messages be given?}
}
\description{
Light wrapper for \code{\link[sbtools]{item_rm_files}}: loops over multiple
sb_id,file pairs if given. For timeseries files, sb_id and file can be
determined using summarize_ts_files(). See examples
}
\examples{
\dontrun{
oops <- summarize_ts_files('dosat_calcGGbts') \%>\% 
  filter(upload_date > as.POSIXct('2016-02-18'), 
    upload_date < as.POSIXct('2016-02-20'), 
    version=='rds') \%>\%
  select(sb_id=ts_item, files=file_name)
mda.streams:::delete_sb_file(oops$sb_id, oops$files)
}
}
\keyword{internal}

