% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config_to_data.R
\name{dev_lubridate_round_date}
\alias{dev_lubridate_round_date}
\title{Round, flour and ceiling methods for date-time objects.}
\usage{
dev_lubridate_round_date(x, unit = c("second", "minute", "hour", "day",
  "week", "month", "year", "quarter"))
}
\arguments{
\item{x}{a vector of date-time objects}

\item{unit}{a character string specifying the time unit to be rounded to.
Should be one of "second", "minute", "hour", "day", "week", "month",
"quarter", or "year."}
}
\value{
x with the appropriate units floored
}
\description{
Copied from https://github.com/hadley/lubridate/blob/master/R/round.r; this
version differs from the CRAN version, which produces NAs on some rounding
tasks. Will be replaced by official lubridate version when the changes are
pushed to CRAN.
}
\keyword{internal}

