% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_meta.R
\name{download_meta}
\alias{download_meta}
\title{Download metadata data to local file destination}
\usage{
download_meta(type, folder = tempdir(), on_remote_missing = c("stop",
  "return_NA"), on_local_exists = c("stop", "skip", "replace"))
}
\arguments{
\item{type}{the metadata type}

\item{folder}{string for a folder location}

\item{on_remote_missing}{character indicating what to do if the}

\item{on_local_exists}{character indicating what to do if the folder already
contains a file with the intended download name}
}
\value{
file handle (character path) for the downloaded file, or NA if the
  timeseries is unavailable on ScienceBase
}
\description{
Download a timeseries file to a user-specified (or temp file)
  location
}
\examples{
\dontrun{
download_meta(type="basic", on_local_exists="replace")
}
}
\author{
Corinna Gries, Jordan S Read, Luke A Winslow, Alison P Appling
}

