% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metab_model.R
\name{download_metab_model}
\alias{download_metab_model}
\title{Download metabolism model .RData to local file}
\usage{
download_metab_model(model_name, folder = tempdir(), version = c("modern",
  "original"), on_remote_missing = c("stop", "return_NA"),
  on_local_exists = c("stop", "skip", "replace"))
}
\arguments{
\item{model_name}{the name of the metab_model file}

\item{folder}{string for a folder location where files should be saved, or a 
vector of as many folders as there are item_ids}

\item{version}{character indicating whether you want the original metab_model
or a modernized one that works with the current streamMetabolizer version}

\item{on_remote_missing}{character indicating what to do if the}

\item{on_local_exists}{character indicating what to do if the folder already 
contains a file with the intended download name}
}
\value{
file handle (character path) for the downloaded file, or NA if the 
  timeseries is unavailable on ScienceBase
}
\description{
WARNING: older metab_models may not be compatible with contemporary 
streamMetabolizer or mda.streams functions. To investigate older models, 
either apply modernize_metab_model to your downloaded files or use 
get_metab_model, which does this for you.
}
\examples{
\dontrun{
download_metab_model("nwis_04087088-200-150730 0.0.7 MLE_for_PRK_wHarvey_and_sw")
}
}
\author{
Alison P Appling
}
