% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metab_run.R
\name{download_metab_run}
\alias{download_metab_run}
\title{Download metabolism run data data to local file destination}
\usage{
download_metab_run(title, files = NA, folder = tempdir(),
  on_remote_missing = c("stop", "return_NA"), on_local_exists = c("stop",
  "skip", "replace"))
}
\arguments{
\item{title}{the title of the metabolism run (date, tag, and strategy 
separated by spaces)}

\item{files}{NA for all files, or a list of vectors of specific filenames 
(character) to download, one vector per title}

\item{folder}{string for a folder location where files should be saved, or a 
vector of as many folders as there are item_ids}

\item{on_remote_missing}{character indicating what to do if the}

\item{on_local_exists}{character indicating what to do if the folder already 
contains a file with the intended download name}
}
\value{
file handle (character path) for the downloaded file, or NA if the 
  timeseries is unavailable on ScienceBase
}
\description{
Download files stored in a metabolism run SB item to a user-specified (or 
temp file) location. A new folder will be created within that location to 
store the downloaded files. The default is to download all available files,
but specific files may be specified with the \code{files} argument.
}
\examples{
\dontrun{
download_ts(var_src = 'doobs_nwis', site_name = 'nwis_06893300')
}
}
\author{
Alison P Appling
}

