% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_site_coords.R
\name{find_site_coords}
\alias{find_site_coords}
\title{Pull the latitude and longitude of a site from ScienceBase}
\usage{
find_site_coords(site_names, format = c("normal", "geoknife"),
  attach.units = (format == "normal"))
}
\arguments{
\item{site_names}{a list of site names such as those returned from
make_site_name()}

\item{format}{character specifying the desired output format. normal is a
human-readable, dplyr-friendly format. geoknife works well for geoknife and
excludes sites with incomplete coordinate data.}

\item{attach.units}{logical. Should units be attached?}
}
\value{
a data.frame of
}
\description{
Pull the latitude and longitude of a site from ScienceBase
}
\examples{
find_site_coords(c("nwis_01467200","styx_001001","nwis_351111089512501"))
find_site_coords(c("nwis_01467200","nwis_09327000","nwis_351111089512501",
    "styx_0001001"), format="geoknife")
}

