% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/find_sites.R
\name{find_sites}
\alias{find_sites}
\title{Higher-level site querying function}
\usage{
find_sites(with_timeseries = NULL, logic = c("all", "any"), ...)
}
\arguments{
\item{with_timeseries}{character vector of timeseries variables (i.e., 1+ of
those listed in get_ts_variables())}

\item{logic}{how to join the constraints in with_timeseries, ...: is any of
the listed parameters sufficient, or do you need all of them to be
available for a site to qualify?}

\item{...}{additional querying arguments yet to be implemented}
}
\value{
a character vector of site IDs
}
\description{
Higher-level site querying function
}
\examples{
\dontrun{
  find_sites()
  find_sites(with_timeseries=c("wtr","doobs"), logic="any")
}
}

