% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_item.R
\name{format_item}
\alias{format_item}
\title{Format an item as a ScienceBase ID or URL, as requested}
\usage{
format_item(item, format)
}
\arguments{
\item{item}{a ScienceBase item data.frame, e.g., as returned from
query_item_identifier}
}
\description{
Internal helper to locate_item. Doesn't check the value of format - that's up
to the calling functions to do, for efficiency.
}
\keyword{internal}

