% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_config.R
\name{get_config}
\alias{get_config}
\title{Download and read a config file from a metab run on SB}
\usage{
get_config(title, config_file = "config.tsv", on_local_exists = "skip")
}
\arguments{
\item{title}{the metab run title. See options using \code{list_metab_runs()}}

\item{config_file}{the name of the config file. It's usually but not always
'config.tsv'; use \code{list_metab_run_files(title)} to check}

\item{on_local_exists}{character indicating what to do if the folder already
contains a file with the intended download name}
}
\description{
Download and read a config file from a metab run on SB
}

