% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metab_data.R
\name{get_metab_data}
\alias{get_metab_data}
\title{Get a data.frame of metabolism input data}
\usage{
get_metab_data(site, sitetime = NA, doobs = NA, dosat = NA, depth = NA,
  wtr = NA, par = NA, disch = NULL, veloc = NULL, sitedate = NULL,
  doinit = NULL, gpp = NULL, er = NULL, K600 = NULL, K600lwr = NULL,
  K600upr = NULL, dischdaily = NULL, velocdaily = NULL, start_date = NA,
  end_date = NA, tag = "0.0.0", strategy = "get_metab_data",
  date = Sys.time(), model = "metab_mle", model_args = "list()",
  simplify = TRUE, verbose = TRUE)
}
\arguments{
\item{simplify}{logical. If there's only one data.frame to retrieve, remove 
it from the enclosing list?}

\item{verbose}{logical. give progress messages?}
}
\value{
a list of data.frames, or if simplify=TRUE and there's only one
  data.frame, a single data.frame of input data
}
\description{
Quick access to metabolism input data (\code{data} arg only)
}
\examples{
\dontrun{
md <- get_metab_data('nwis_08062500')
md <- get_metab_data('styx_004001')
login_sb()
md <- get_metab_data('styx_004001', doobs=choose_data_source('doobs','styx_004001','sim',type='pred',src='160411 0.0.18'))
}
}

