% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metab_prep.R
\name{get_metab_prep}
\alias{get_metab_prep}
\title{Get a data.frame of metabolism input data}
\usage{
get_metab_prep(site, out = c("data", "data_daily", "info"), sitetime = NA,
  doobs = NA, dosat = NA, depth = NA, wtr = NA, par = NA,
  disch = NULL, veloc = NULL, sitedate = NULL, doinit = NULL,
  gpp = NULL, er = NULL, K600 = NULL, K600lwr = NULL, K600upr = NULL,
  dischdaily = NULL, velocdaily = NULL, start_date = NA, end_date = NA,
  tag = "0.0.0", strategy = "get_metab_data", date = Sys.time(),
  model = "metab_mle", model_args = "list()", simplify_config = TRUE,
  simplify_out = TRUE, verbose = TRUE)
}
\arguments{
\item{simplify_config}{logical. If the implied config file has only one row, should the list be reduced to one element?}

\item{simplify_out}{If \code{out} has only one element, should the list of lists be reduced to a list of data.frames?}

\item{verbose}{logical. give progress messages?}
}
\description{
Quick-get function for manual getting of metabolism data
}
\examples{
md <- get_metab_prep('nwis_08062500') # list of $data, $data_daily, $info
md <- get_metab_prep('nwis_08062500', simplify_config=FALSE) # list of list of $data, $data_daily, $info
md <- get_metab_prep('nwis_08062500', out='data') # single data.frame
md <- get_metab_prep(c('nwis_08062500','nwis_01646500'), out='data') # list of data.frames
}

