% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_nldas_df.R
\name{get_nldas_df}
\alias{get_nldas_df}
\title{get nwis data as data.frame}
\usage{
get_nldas_df(site, variable_name, das_code, start_date, end_date)
}
\arguments{
\item{site}{valid NWIS site ID}

\item{variable_name}{short name of variable \code{\link{get_ts_variables}}}

\item{das_code}{NLDAS variable name (see \code{\link[geoknife]{getDataIDs}})}

\item{...}{additional arguments passed to geoknife, e.g., startDate}
}
\value{
a data.frame of timeseries data from NWIS, or NULL if no data exist
}
\description{
get data from nwis and return as create data.frame
}
\examples{
\dontrun{
df <- get_nldas_df(site = "06893820", variable_name = "baro", das_code = "pressfc",
                 start_date = '2014-01-01', end_date = '2014-02-01')
}
}

