% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_nwis_df}
\alias{get_nwis_df}
\title{get nwis data as data.frame}
\usage{
get_nwis_df(site, variable_name, p_code, ...)
}
\arguments{
\item{site}{valid NWIS site ID}

\item{variable_name}{short name of variable \code{\link{get_ts_variables}}}

\item{p_code}{NWIS parameter code}

\item{...}{additional arguments passed to \code{\link{readNWISuv}}}
}
\value{
a data.frame of timeseries data from NWIS, or NULL if no data exist
}
\description{
get data from nwis and return as create data.frame
}
\examples{
df <- get_nwis_df(site = "06893820", variable_name = "doobs", p_code = "00300",
                 startDate = '2014-01-01', endDate = '2014-02-01')
df <- get_nwis_df(site = "nwis_06893820", variable_name = "doobs", p_code = "00300",
                 startDate = '2014-01-01', endDate = '2014-02-01')
}

