% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scheme.R
\name{get_scheme}
\alias{get_scheme}
\title{Get the scheme name for this project}
\usage{
get_scheme(warn = TRUE)
}
\arguments{
\item{warn}{logical specifying whether a warning should be given if the
scheme is not set}
}
\value{
character name of the scheme
}
\description{
Gives a warning and returns NULL if scheme has not been set
}

