% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_src_codes.R
\name{get_src_codes}
\alias{get_src_codes}
\title{Get source codes used in this project}
\usage{
get_src_codes(src, out = names(src_codes), type = c("data", "calc"),
  drop = (length(out) == 1), use_names)
}
\arguments{
\item{src}{a valid source name for project (e.g., 'nwis')}

\item{out}{character or character vector of name types to include in output.
If there is only one, the result will be a vector/value. One or more of the
given options may be selected}

\item{type}{optional filter criterion: setting type to a subset of the
default options will restrict the output to variables that have one of
those selected variable types.}

\item{drop}{logical. In data.frame subsetting, should a single column be
dropped to a vector or a single row be dropped to a list?}

\item{use_names}{logical. Should names be attached to the data.frame rows or
vector elements? The default tries to be smart about what would be
convenient.}
}
\value{
vector or data.frame of the requested variables and outnames.
}
\description{
A lookup for parameter codes and long_names based on the var
  convention used in mda.streams
}
\examples{
get_src_codes()
get_src_codes('nwis')
get_src_codes(out=c('r_function'), type='calc')
}

