% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_var_codes.R
\name{get_var_codes}
\alias{get_var_codes}
\title{get project codes}
\usage{
get_var_codes(shortname, with.longname = FALSE)
}
\arguments{
\item{shortname}{a valid shortname for project (e.g., 'disch')}

\item{with.longname}{boolean for whether to include longnames in output}
}
\description{
get project codes
}
\examples{
get_var_codes('wtr')
get_var_codes('disch', TRUE)
}

