% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_var_src_codes.R
\name{get_var_src_codes}
\alias{get_var_src_codes}
\title{Retrieve the var and src codes for mda.streams}
\usage{
get_var_src_codes(..., out, drop = TRUE)
}
\arguments{
\item{out}{a list of arguments to pass to \code{dplyr::select_}}

\item{filter_by}{a list of arguments to pass to \code{dplyr::filter_}}
}
\description{
Facilitates quick filtering and selecting
}
\examples{
get_var_src_codes()
get_var_src_codes(var=="baro", out='var')
get_var_src_codes(var=="baro", out='var', drop=FALSE)
get_var_src_codes(var=="baro", out=c('var','var_descrip'))
get_var_src_codes(var=="baro", out=c('var','var_descrip'), drop=FALSE)
get_var_src_codes(out=c('var','var_src','p_code'))
get_var_src_codes(src_type=="data", !is.na(p_code),
  out=list('var','var_src','p_code', 'src_type'))
}

