% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_watershed_WFS.R
\name{get_watershed_WFS}
\alias{get_watershed_WFS}
\title{get WFS from site}
\usage{
get_watershed_WFS(site)
}
\arguments{
\item{site}{a site identifier}
}
\value{
a character string for WFS url or NULL if watershed doesn't exist for site, or site doesn't exist.
}
\description{
get a WFS url from a site that has a 'watershed'
}
\examples{

\dontrun{
site <- "nwis_04165500"
get_watershed_WFS(site)

# will fail
get_watershed_WFS('fake_site'))
}

}
\seealso{
\code{\link{get_watershed_WMS}}
}

