% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/init_nwis_sites.R
\name{init_nwis_sites}
\alias{init_nwis_sites}
\title{generate nwis site_ids for given p_codes}
\usage{
init_nwis_sites(p_codes, state_codes = "all")
}
\arguments{
\item{p_codes}{a character vector of NWIS p_codes}

\item{state_codes}{a character vector of state codes
(e.g., \code{c("WA","WI")}). Default is all, which will use all state
codes available to \code{\link{get_state()}}. Note this includes Puerto Rico and more.}
}
\value{
a character vector of NWIS sites, appended with 'nwis_'
}
\description{
finds all NWIS sites that meet given data requirements
}
\examples{
\dontrun{
init_nwis_sites(p_codes = c('00010', '00060', '00095', '00300'))
}
}

