% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Get a vector of timeseries dataset names}
\usage{
list_datasets(site_name, type = c("ts", "watershed"), ...)
}
\arguments{
\item{type}{character. one or more dataset types to return}

\item{...}{additional arguments passed to
  \code{\link[sbtools]{session_check_reauth}}, for example \code{username}}

\item{site_names}{a list of site names such as those returned from
make_site_name()}
}
\value{
an alphabetically sorted character vector of unique timeseries
  variable names for given sites
}
\description{
list_datasets(site) returns the data available to a specific site. In
contrast, get_var_codes() returns a list of all possible variables and
get_var_codes(type="ts") returns all possible timeseries variables.
}
\examples{
\dontrun{
list_datasets(site_name = 'nwis_295826095082200')
}
}

