% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_metab_models.R
\name{list_metab_models}
\alias{list_metab_models}
\title{List the available metab_model objects}
\usage{
list_metab_models()
}
\value{
a character vector of titles of the metab_model .RData files posted on SB
}
\description{
List the available metab_model objects
}

