% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_metab_runs.R
\name{list_metab_run_files}
\alias{list_metab_run_files}
\title{List the files stored in a metabolism runs}
\usage{
list_metab_run_files(title, out = c("filename", "size_bytes", "url"))
}
\arguments{
\item{title}{a metab_run title}

\item{out}{the columns to return}
}
\value{
dataframe of file names and sizes within a metab_run
}
\description{
List the files stored in a metabolism runs
}
