% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_metab_runs.R
\name{list_metab_runs}
\alias{list_metab_runs}
\title{List the available metabolism runs}
\usage{
list_metab_runs()
}
\value{
a character vector of titles of the metab_runs posted on SB
}
\description{
List the available metabolism runs
}
