% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_metas.R
\name{list_metas}
\alias{list_metas}
\title{List the available metadata tables}
\usage{
list_metas()
}
\value{
a character vector of titles of the meta tables posted on SB
}
\description{
List the available metadata tables
}

