% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_tses.R
\name{list_tses}
\alias{list_tses}
\title{List the timeseries data available for a particular site}
\usage{
list_tses(site_name)
}
\arguments{
\item{site_name}{a valid powstreams site name (see \link{list_sites})}
}
\value{
a character vector of timeseries names, or NULL if none exists.
}
\description{
List the timeseries data available for a particular site
}
\examples{
list_tses(site = "nwis_50231500")
}
\seealso{
\code{\link{download_ts}}
}

