% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_manual.R
\name{load_google_sheet}
\alias{load_google_sheet}
\title{Read in data from a Google Spreadsheet}
\usage{
load_google_sheet(sheets_key, na.string = "", header = TRUE)
}
\arguments{
\item{sheets_key}{the Google document key, a string of letters and numbers found in the browser URL after /d/}

\item{na.string}{the string to convert to NA if it's all that's in a cell}

\item{header}{logical. Expect a header row?}
}
\description{
Data-reading function from comments in
http://blog.revolutionanalytics.com/2014/06/reading-data-from-the-new-version-of-google-spreadsheets.html
}
\keyword{internal}
