% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_folder.R
\name{locate_folder}
\alias{locate_folder}
\title{Find a high-level folder on ScienceBase}
\usage{
locate_folder(folder = c("project", "metab_runs", "metab_models", "sites",
  "sites_meta", "ts_meta", "ideas", "presentations", "proposals",
  "publications"), format = c("id", "url"), by = c("tag", "dir", "either"),
  limit = 5000, browser = (format == "url"))
}
\arguments{
\item{folder}{the folder to locate}

\item{format}{character indicating whether the folder should be returned as 
an ID or as a full URL}

\item{by}{character indicating how to search for the item: using tags ("tag",
the default and recommended option), by scanning the parent directory for 
the desired title ("dir"), or both in combination ("either")?}

\item{limit}{number of items to return, as in 
\code{\link[sbtools]{query_item_identifier}} or 
\code{\link[sbtools]{item_list_children}}}

\item{browser}{logical. Should the URL be opened in a browser?}
}
\description{
Find a high-level folder on ScienceBase
}
\examples{
\dontrun{
locate_folder("publications", format="url")
testthat::expect_error(locate_folder("cvs", format="url"))
}
}
