% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_item.R
\name{locate_item}
\alias{locate_item}
\title{Find a folder or other item on ScienceBase}
\usage{
locate_item(key = c("nwis_02322688", "project", "presentations", "proposals",
  "publications", "sites"), type = c("site_root", "root"), format = c("id",
  "item_url", "folder_url"), ...)
}
\arguments{
\item{key}{the item key, probably identical to the folder name}

\item{type}{the item type, probably either root or site_root}

\item{format}{character indicating whether the folder should be returned as
an ID or as a full URL}

\item{...}{other args passed to \code{\link[sbtools]{session_check_reauth}}}
}
\description{
Find a folder or other item on ScienceBase
}
\examples{
\dontrun{
mda.streams:::locate_item(key="sites", type="root")
mda.streams:::locate_item(key=c("sites","proposals","publications"),
  type="root", format="folder_url")
mda.streams:::locate_item(key="nwis_02322688", type="site_root")
}
}

