% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_item.R
\name{locate_item_by_dir}
\alias{locate_item_by_dir}
\title{Locate items using the folder structure rather than tags}
\usage{
locate_item_by_dir(parent, title, format = c("id", "item_url", "folder_url"),
  ...)
}
\arguments{
\item{parent}{the ScienceBase ID of the parent whose children to search}

\item{title}{the title to seek among the children}
}
\value{
an id, formatted as requested
}
\description{
This function is useful for ScienceBase directory maintenance, because it
helps locate items that have no tags along with those that do. Items without
tags are created on post_ts operations that fail partway through.
}
\keyword{internal}

