% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_metab_run.R
\name{locate_metab_run}
\alias{locate_metab_run}
\title{Find a metab_run item on ScienceBase}
\usage{
locate_metab_run(date, tag, strategy, title = paste0(date, " ", tag, " ",
  strategy), format = c("id", "url"), by = c("tag", "dir", "either"),
  limit = 5000, browser = (format == "url"))
}
\arguments{
\item{date}{the date as a character string in \%Y\%m\%d format}

\item{tag}{the tag as a character string}

\item{strategy}{the run strategy as a character string}

\item{title}{the title of the metabolism modeling run you want. this will be
automatically constructed if date, tag, and strategy are available; if any
of those is missing, title should be supplied.}

\item{format}{character indicating whether the folder should be returned as
an ID or as a full URL}

\item{by}{character indicating how to search for the item: using tags ("tag",
the default and recommended option), by scanning the parent directory for
the desired title ("dir"), or both in combination ("either")?}

\item{limit}{number of items to return, as in
\code{\link[sbtools]{query_item_identifier}} or
\code{\link[sbtools]{item_list_children}}}

\item{browser}{logical. Should the URL be opened in a browser?}
}
\description{
Find a metab_run item by title
}
\examples{
\dontrun{
locate_metab_run('150714 0.0.2 local_makefile_run')
}
}

