% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_ts.R
\name{locate_ts}
\alias{locate_ts}
\title{Identify a time series from a variable name, data type, and specific data
source (i.e., as specified in a metab model config file)}
\usage{
locate_ts(site, var, type, src)
}
\arguments{
\item{var}{character. The variable name, e.g. "doobs", "disch", or "wtr"}

\item{type}{character. The source type (e.g., "local", "proxy", "model")}

\item{src}{character. The specific source (e.g., "nwis_01474000", "bird")}
}
\value{
URL pointer to a ScienceBase item

locate_ts("nwis_13206400", )
}
\description{
Identify a time series from a variable name, data type, and specific data
source (i.e., as specified in a metab model config file)
}

