% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_ts_meta.R
\name{locate_ts_meta}
\alias{locate_ts_meta}
\title{Find a timeseries metadata object on ScienceBase}
\usage{
locate_ts_meta(type, format = c("id", "url"), by = c("tag", "dir",
  "either"), limit = 5000, browser = (format == "url"))
}
\arguments{
\item{type}{the type of ts metadata you want (currently there's only var_src_codes)}

\item{format}{character indicating whether the folder should be returned as 
an ID or as a full URL}

\item{by}{character indicating how to search for the item: using tags ("tag",
the default and recommended option), by scanning the parent directory for 
the desired title ("dir"), or both in combination ("either")?}

\item{limit}{number of items to return, as in 
\code{\link[sbtools]{query_item_identifier}} or 
\code{\link[sbtools]{item_list_children}}}

\item{browser}{logical. Should the URL be opened in a browser?}
}
\description{
Find a timeseries metadata object on ScienceBase
}
\examples{
\dontrun{
locate_ts_meta('basic')
}
}

