% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_archive_name}
\alias{make_archive_name}
\title{Create a archive file name for a file on ScienceBase}
\usage{
make_archive_name(filename, file_field)
}
\arguments{
\item{filename}{of the file that is to be archived}

\item{the}{file field of the ScienceBase item that will be archived (list)}
}
\value{
a file name to be used as an archive
}
\description{
Create a archive file name for a file on ScienceBase
}

