% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_meta_path}
\alias{make_meta_path}
\title{Create a standardized file name for specific file contents}
\usage{
make_meta_path(type = c("basic", "metabinput", "dvqcoefs"), folder)
}
\arguments{
\item{type}{the type of metadata the file will include}

\item{folder}{the folder to write the file in, or missing}
}
\value{
a full file path
}
\description{
Create a standardized file name for specific file contents
}

