% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_metab_config.R
\name{make_metab_config}
\alias{make_metab_config}
\title{Make a metabolism config file}
\usage{
make_metab_config(site, sitetime = NA, doobs = NA, dosat = NA,
  depth = NA, wtr = NA, par = NA, disch = NULL, veloc = NULL,
  sitedate = NULL, doinit = NULL, gpp = NULL, er = NULL, K600 = NULL,
  K600lwr = NULL, K600upr = NULL, dischdaily = NULL, velocdaily = NULL,
  start_date = NA, end_date = NA, tag = "0.0.0",
  strategy = "get_metab_data", date = Sys.time(), model = "metab_mle",
  model_args = "list()")
}
\arguments{
\item{site}{site names}

\item{sitetime}{Data Source for mean solar time. See Data Source Format 
below.}

\item{doobs}{Data Source for dissolved oxygen concentrations. See Data Source
Format below.}

\item{dosat}{Data Source for dissolved oxygen saturation concentrations. See 
Data Source Format below.}

\item{depth}{Data Source for mean stream depth. See Data Source Format below.}

\item{wtr}{Data Source for water temperature. See Data Source Format below.}

\item{par}{Data Source for light (photosynthetically available radiation, 
PAR). See Data Source Format below.}

\item{disch}{Data Source for unit-value stream discharge, for use in 
identifying daily priors or fixed values for K600. See Data Source Format 
below.}

\item{veloc}{Data Source for unit-value flow velocity, for use in identifying
daily priors or fixed values for K600. See Data Source Format below.}

\item{sitedate}{Data Source for the dates of interest. See Data Source Format
below.}

\item{doinit}{Data Source for the first DO observation on each date to model,
for use in data simulation. See Data Source Format below.}

\item{gpp}{Data Source for daily gross primary productivity rates for use in 
data simulation. See Data Source Format below.}

\item{er}{Data Source for ecosystem respiration rates for use in data 
simulation. See Data Source Format below.}

\item{K600}{Data Source for reaeration rates for use in data simulation. See 
Data Source Format below.}

\item{K600lwr}{Data Source for lower bound on reaeration rates for use in
data simulation. See Data Source Format below.}

\item{K600upr}{Data Source for upper bound on reaeration rates for use in
data simulation. See Data Source Format below.}

\item{dischdaily}{Data Source for daily mean stream discharge, for use in 
identifying daily priors or fixed values for K600. See Data Source Format 
below.}

\item{velocdaily}{Data Source for daily mean flow velocity, for use in 
identifying daily priors or fixed values for K600. See Data Source Format 
below.}

\item{start_date}{NA or datetime, to be coerced with 
\code{as.POSIXct(start_date, tz="UTC")}, at which to start the data passed 
to the metab model}

\item{end_date}{NA or datetime, to be coerced with \code{as.POSIXct(end_date,
tz="UTC")}, at which to end the data passed to the metab model}

\item{tag}{character of form "1.0.2" that uniquely identifies this set of 
modeling runs.}

\item{strategy}{character, or vector of length sites, describing this set of 
modeling runs in concise English.}

\item{date}{POSIXct indicating the date of config construction. It is 
strongly recommended to use the default.}

\item{model}{character. the name of the metabolism model to construct}

\item{model_args}{character, in R language, specifying any arguments to pass 
to the model function}

\item{var}{a variable, passed in as a named object, whose name will be used
as the variable name and whose value will be used (1) as the output of 
choose_data_source if a data.frame, (2) as a request for an unused var if
NULL, (3) as a request for a priority local selection if NA, and (4) as a
specific src value if character}
}
\description{
Quick-make function for manual making of a model config file
}
\section{Data Source Format}{

  
  If a data source is given as NA, the 'priority local' timeseries option 
  will be selected.
  
  If a data source is given as a character, the character will be used to 
  specify the \code{src} argument to \code{choose_data_source} with 
  code{type='ts'}.
  
  If a data source is given as NULL, the variable will be set to 'unused
  var'.
  
  For the most flexibility, each parameter whose definition begins with Data 
  Source may be supplied as a 4-column data.frame with column names 
  c('type','site','src','logic'). These specify where to find the data for a 
  given variable. The easiest way to create such a data.frame is usually with
  \code{\link{choose_data_source}}, though it may also be created manually.
  
  The variables requiring Data Source specification, along with their 
  expected units, are defined in the help file for \code{\link{mm_data}}.
}
\examples{
make_metab_config('nwis_08062500')
make_metab_config('nwis_08062500', sitetime='calcLon', doobs='nwis', 
  dosat='calcGGbts', depth='calcDischHarvey', wtr='nwis', par='calcSw')
helper function for translating the above simpler args into args suitable 
for stage_metab_config

}

