% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_metab_model_name}
\alias{make_metab_model_name}
\title{Produce a name unique to a single metab_model}
\usage{
make_metab_model_name(title, row, site)
}
\arguments{
\item{title}{metab_run title}

\item{row}{the config row number == metab_all.RData element number}

\item{site}{the site name}
}
\description{
Produce a name unique to a single metab_model
}

