% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_metab_run_title}
\alias{make_metab_run_title}
\title{Combine the date, tag, and strategy of a model_run into a title}
\usage{
make_metab_run_title(date, tag, strategy)
}
\arguments{
\item{date}{text date, e.g. 070802}

\item{tag}{text tag, e.g., 0.0.3}

\item{strategy}{text strategy, e.g., nighttime_k}
}
\description{
Combine the date, tag, and strategy of a model_run into a title
}
