% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_site_name}
\alias{make_site_name}
\title{Translate a site name from database and site number to a
ScienceBase+mda.streams site ID}
\usage{
make_site_name(database = c("nwis", "nldas"), sitenum)
}
\arguments{
\item{database}{a character or character vector of databases, probably in
\code{c("nwis","nldas")}.}

\item{sitenum}{integer or character, coercible to character, representing the
site code as used by the database.}
}
\value{
site ID in ScienceBase and mda.streams lingo
}
\description{
Translate a site name from database and site number to a
ScienceBase+mda.streams site ID
}

