% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_ts_name}
\alias{make_ts_name}
\title{Translate a timeseries name from mda.streams to ScienceBase}
\usage{
make_ts_name(var_var_src, src)
}
\arguments{
\item{var_var_src}{may be either a variable, e.g. "doobs", or a var_src, e.g.
"doobs_nwis".}

\item{src}{a data source; optional if var_var_src is specified with both
variable and source}
}
\value{
timseries name[s] in ScienceBase lingo (e.g., "\code{ts_wtr})
}
\description{
Translate a timeseries name from mda.streams to ScienceBase
}

