% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_ts_variable}
\alias{make_ts_variable}
\title{create timeseries pointer}
\usage{
make_ts_variable(variable)
}
\arguments{
\item{variable}{a timeseries name (e.g., \code{wtr})}
}
\value{
a timseries string
}
\description{
create timeseries pointer
}

