% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_var_src}
\alias{make_var_src}
\title{Translate a variable and data source into a var_src}
\usage{
make_var_src(var, src)
}
\arguments{
\item{var}{a variable shortname from get_var_codes(out="shortname"), e.g.,
"doobs"}

\item{src}{a data source from the comma-separated lists in
get_var_codes(out="sources")}
}
\description{
Translate a variable and data source into a var_src
}

