% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_meta_path}
\alias{parse_meta_path}
\title{Split a metadata file path into contents}
\usage{
parse_meta_path(file_path, out = c("dir_name", "file_name", "type",
  "meta_type"), use_names = length(out) > 1)
}
\arguments{
\item{file_path}{the path[s] to split}

\item{out}{the columns to return}

\item{use_names}{logical. Should names be attached to the data.frame rows or
list elements?}
}
\value{
a data.frame, one row per path
}
\description{
Split a metadata file path into contents
}

