% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_metab_model_name}
\alias{parse_metab_model_name}
\title{Parse a metab_model name}
\usage{
parse_metab_model_name(model_name, out = c("title", "row", "site", "date",
  "tag", "strategy"), use_names = length(model_name) > 1)
}
\arguments{
\item{model_name}{the unique name of the metab_model as produced by 
make_metab_model_name}

\item{out}{the columns to return}

\item{use_names}{logical. Should names be attached to the data.frame rows or 
list elements?}
}
\description{
Parse a metab_model name
}

