% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_metab_model_path}
\alias{parse_metab_model_path}
\title{Parse a file path into metab_model information}
\usage{
parse_metab_model_path(file_path, out = c("dir_name", "file_name",
  "model_name", "title", "row", "site", "date", "tag", "strategy"),
  use_names = length(file_path) > 1)
}
\arguments{
\item{file_path}{the file path}

\item{out}{character. which columns to return}

\item{use_names}{attach row/vector names?}
}
\description{
Parse a file path into metab_model information
}

