% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_metab_run_title}
\alias{parse_metab_run_title}
\title{Parse the title of a metab_model run}
\usage{
parse_metab_run_title(title, out = c("date", "tag", "strategy"),
  use_names = length(title) > 1)
}
\arguments{
\item{title}{the model run title}

\item{out}{the columns to return}

\item{use_names}{logical. Should names be attached to the data.frame rows or
list elements?}
}
\value{
a data.frame, one row per path
}
\description{
Parse the title of a metab_model run
}

