% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_site_name}
\alias{parse_site_name}
\title{Split site name into siteID (used for NWIS site numbers)}
\usage{
parse_site_name(site_name, out = "sitenum", use_names = length(out) > 1)
}
\arguments{
\item{site_name}{a valid ScienceBase site (e.g., "nwis_0338102")}

\item{out}{character, length 1 or 2 selected from 
\code{c("database","sitenum")} indicating whether you want to get back the 
database, the site number, or both.}

\item{use_names}{logical. Should the return vector be named according to the
input values?}
}
\value{
the database, sitenum, or both. If both, the return value is a 
  data.frame; otherwise it's a vector.
}
\description{
Split site name into siteID (used for NWIS site numbers)
}

