% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_ts_archive_name}
\alias{parse_ts_archive_name}
\title{Split archived file path into its contents}
\usage{
parse_ts_archive_name(file_name, out = c("ts_name", "is_archive",
  "creation_date"), use_names = length(file_name) > 1)
}
\arguments{
\item{file_name}{a valid file basename for ts file}

\item{out}{a character for desired output ('dir_name','file_name', 
'site_name', 'ts_name', 'creation_date' or any of the out names from
parse_ts_name or parse_site_name)}

\item{use_names}{logical. Should the return vector be named according to the 
input values?}
}
\value{
a character
}
\description{
Subsidiary to parse_ts_path, which is the exported function
}
\examples{
mda.streams:::parse_ts_archive_name(
  c('ARCHIVE_20160324120400-nwis_05406479-ts_doobs_nwis.tsv.gz', 
    'nwis_05406480-ts_doobs_nwis.tsv.gz'), use_names=FALSE)
}

