% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_ts_archive_path}
\alias{parse_ts_archive_path}
\title{Split archived file path into contents}
\usage{
parse_ts_archive_path(file_path, out = c("dir_name", "file_name", "version",
  "site_name", "ts_name", "var_src", "var", "src", "database", "sitenum",
  "creation_date"), use_names = length(file_path) > 1)
}
\arguments{
\item{file_path}{a valid file path for ts file}

\item{out}{a character for desired output ('dir_name','file_name',
'site_name', 'ts_name', 'creation_date' or any of the out names from parse_ts_name or
parse_site_name)}

\item{use_names}{logical. Should the return vector be named according to the 
input values?}
}
\value{
a character
}
\description{
Split archived file path into contents
}
\examples{
mda.streams:::parse_ts_archive_path(c('ARCHIVE_20160324_nwis_05406479-ts_doobs_nwis.tsv.gz', 
'ARCHIVE_20160324_nwis_0406479-ts_doobs_nwis.tsv.gz'), use_names = T)
}

