% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_var_src}
\alias{parse_var_src}
\title{Translate a var_src into a variable and a source}
\usage{
parse_var_src(var_src, out = c("var", "src"), use_names = FALSE)
}
\arguments{
\item{var_src}{an underscore-separated combination of a variable and its data
source}

\item{out}{the parsed fields to return.}

\item{use_names}{logical. Should names/rownames be attached to the
vector/data.frame?}
}
\value{
if length(out)==1 a vector, else a data.frame
}
\description{
Translate a var_src into a variable and a source
}

