% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_indy_site.R
\name{post_indy_site}
\alias{post_indy_site}
\title{Create a indy site and post a list of timeseries files to it}
\usage{
post_indy_site(file_list)
}
\arguments{
\item{file_list}{a list of files as given by stage_indy_site. The site name
will be determined from the first item (a metadata vector) in this list.}
}
\description{
This function is easiest to use when applied to the output of
\code{\link{stage_indy_site}}.
}
\examples{
\dontrun{
staged <- stage_indy_site(info="help file example for post_indy_site")
login_sb()
post_indy_site(staged)
}
}

