% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_meta.R
\name{post_meta}
\alias{post_meta}
\title{Post new metadata file[s] to SB}
\usage{
post_meta(files, on_exists = c("stop", "skip", "replace", "merge"),
  verbose = TRUE)
}
\arguments{
\item{files}{character one or more files to post to SB}

\item{on_exists}{character. what should be done when an item already exists?}

\item{verbose}{logical. Should status messages be given?}
}
\value{
an item list
}
\description{
If a file with the same name already exists in Sites_meta, the
\code{on_exists} argument specifies how to manage the conflict.
}
\examples{
\dontrun{
login_sb()
set_scheme("mda_streams_dev")
metafile <- stage_meta(sites=list_sites(), folder="temp")
post_meta(metafile, on_exists="stop", verbose=TRUE)
set_scheme("mda_streams")
}
}
\author{
Alison Appling
}
