% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_metab_model.R
\name{post_metab_model}
\alias{post_metab_model}
\title{Post individual metab_model object[s] to SB}
\usage{
post_metab_model(files, on_exists = c("stop", "skip", "replace_file"),
  verbose = TRUE)
}
\arguments{
\item{files}{character one or more files to post to SB}

\item{on_exists}{character. what should be done when an item already exists?}

\item{verbose}{logical. Should status messages be given?}
}
\value{
an item list
}
\description{
Post individual metab_model object[s] to SB
}
\examples{
\dontrun{
login_sb()
set_scheme("mda_streams_dev")
# stage_metab_model...
# post_metab_model...
set_scheme("mda_streams")
}
}
\author{
Alison Appling
}

