% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_metab_run.R
\name{post_metab_run}
\alias{post_metab_run}
\title{Post files to a model run item on SB}
\usage{
post_metab_run(folder, files, on_exists = c("stop", "skip", "addfiles"),
  verbose = TRUE)
}
\arguments{
\item{folder}{a single folder name from which the tag, strategy, and date
(and therefore metab_run title) will be pulled, and where named files will
be sought}

\item{files}{a vector of filenames to select from within folder and to post
to the new SB item}

\item{on_exists}{character. what should be done when an item already exists?}

\item{verbose}{logical. Should status messages be given?}
}
\description{
Post any number of files to an SB item representing a complete, multi-site
model run. These files will probably include a .tsv config file, an .RData
file containing a list of model fit objects, and an .Rout file containing the
log information from the run.
}

