% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_site.R
\name{post_site}
\alias{post_site}
\title{Create new site in SB}
\usage{
post_site(sites, replace_existing = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{sites}{one or more local site IDs (e.g. nwis_09238475)}

\item{replace_existing}{logical. Should an item that already exists be
replaced?}

\item{verbose}{logical. Should status messages be given?}

\item{...}{Additional parameters supplied to \code{\link[sbtools]{session_check_reauth}}}
}
\value{
an item list
}
\description{
Create new site in SB
}
\examples{
\dontrun{
locate_site(c("nwis_08437710","nwis_03401385","nwis_02423397"), format="url")
post_site("nwis_08437710") # returns NA
post_site("nwis_08437710", replace_existing=TRUE)
post_site(c("nwis_08437710","nwis_03401385","nwis_02423397"), replace_existing=TRUE)
}
}
\author{
Corinna Gries
}

