% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_styx_site.R
\name{post_styx_site}
\alias{post_styx_site}
\title{Create a styx site and post a list of timeseries files to it}
\usage{
post_styx_site(file_list)
}
\arguments{
\item{file_list}{a list of files as given by stage_styx_site. The site name 
will be determined from the first item (a metadata vector) in this list.}
}
\description{
This function is easiest to use when applied to the output of 
\code{\link{stage_styx_site}}.
}
\examples{
\dontrun{
staged <- stage_styx_site(info="help file example for post_styx_site")
login_sb()
post_styx_site(staged)
}
}

