% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_ts.R
\name{post_ts}
\alias{post_ts}
\title{Post a new timeseries file to SB}
\usage{
post_ts(files, on_exists = c("stop", "skip", "replace", "merge"),
  verbose = TRUE)
}
\arguments{
\item{files}{a string vector of file paths for POSTing}

\item{on_exists}{character specifying an action when a file to post already
exists on ScienceBase}

\item{verbose}{logical. Should status messages be given?}
}
\description{
Post a staged file from the local computer to ScienceBase
}
\examples{
\dontrun{
files <- stage_nldas_ts(sites = c("nwis_06893820","nwis_01484680"), var = "baro",
                        times = c('2014-01-01 00:00','2014-01-01 05:00'))
post_ts(files)
}
}
\author{
Luke Winslow, Corinna Gries, Jordan S Read
}

