% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post_watershed}
\alias{post_watershed}
\title{Post a new watershed shape to a site on SB}
\usage{
post_watershed(site, files, session)
}
\arguments{
\item{site}{The unique site identifier}

\item{files}{String vector of files that make up the shapefile}

\item{session}{A SB session from \link{authenticate_sb}}
}
\description{

}
\author{
Luke Winslow
}

