% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Read a config file}
\usage{
read_config(file)
}
\arguments{
\item{file}{a config filename or path/filename}
}
\value{
config data.frame
}
\description{
Read a config file
}

