% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_meta.R
\name{read_meta}
\alias{read_meta}
\title{Read metadata file for mda.streams into data.frame}
\usage{
read_meta(file)
}
\arguments{
\item{file}{a valid ts file path}
}
\value{
a metadata data.frame
}
\description{
Read metadata file for mda.streams into data.frame
}
\examples{
\dontrun{
file <- download_meta(type='basic')
meta_basic <- read_meta(file)
}
}
\author{
Alison P. Appling
}
\seealso{
\code{\link{download_meta}}
}

