% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_ts.R
\name{read_ts}
\alias{read_ts}
\title{read ts format (timeseries) for mda.streams into data.frame}
\usage{
read_ts(file)
}
\arguments{
\item{file_handle}{a valid ts file handle}
}
\value{
a timeseries data.frame
}
\description{
read ts format (timeseries) for mda.streams into data.frame
}
\examples{
\dontrun{
file <- download_ts(site = "nwis_06893820", variable = "baro", overwrite_file = TRUE)
dissolved_oxygen <- read_ts(file)
}
}
\author{
Jordan S. Read
}

