% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_meta.R
\name{repair_meta}
\alias{repair_meta}
\title{Repair a meta item that is missing its identifier tags}
\usage{
repair_meta(type, limit = 5000)
}
\arguments{
\item{type}{one or more meta types, e.g., "basic"}

\item{limit}{the maximum number of items to return in the SB query to find 
all listed var_src:site_name combinations}
}
\description{
Sometimes ts items/files get posted but the identifiers don't get updated, 
making it harder to search for the item. This function gives ScienceBase 
another chance to add tags.
}
\examples{
\dontrun{
repair_meta("basic")
}
}

