% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_metab_model.R
\name{repair_metab_model}
\alias{repair_metab_model}
\title{Repair a metab_model item that is missing its identifier tags}
\usage{
repair_metab_model(model_name, verbose = FALSE, limit = 5000)
}
\arguments{
\item{model_name}{one or more model_name strings, e.g., "nwis_05515500-35-150729 0.1.2 compare_par_srces"}

\item{verbose}{should status messages be given?}

\item{limit}{the maximum number of items to return in the SB query}
}
\description{
Sometimes items/files get posted but the identifiers don't get updated, 
making it harder to search for the ts item. This function gives ScienceBase 
another chance to add tags.
}
