% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_metab_run.R
\name{repair_metab_run}
\alias{repair_metab_run}
\title{Repair a metab_run item that is missing its identifier tags}
\usage{
repair_metab_run(title, limit = 5000)
}
\arguments{
\item{title}{one or more metab titles, e.g., "150714 0.0.2 local_makefile_run"}

\item{limit}{the maximum number of items to return in the SB query to find 
all listed var_src:site_name combinations}
}
\description{
Sometimes metab_run items/files get posted but the identifiers don't get
updated, making it harder to search for the item. This function gives
ScienceBase another chance to add tags.
}
\examples{
\dontrun{
repair_meta("basic")
}
}
