% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_item.R
\name{repair_ts}
\alias{repair_ts}
\title{Repair a ts item that is missing its identifier tags}
\usage{
repair_ts(var_src, site_name, limit = 5000)
}
\arguments{
\item{var_src}{one or more var_src strings, e.g., "doobs_nwis"}

\item{site_name}{one or more site names, e.g., "nwis_040871488"}

\item{limit}{the maximum number of items to return in the SB query to find
  all listed var_src:site_name combinations}
}
\description{
Sometimes ts items/files get posted but the identifiers don't get updated,
making it harder to search for the ts item. This function gives ScienceBase
another chance to add tags.
}

