% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_sb.R
\name{sb_renew_login}
\alias{sb_renew_login}
\title{If needed, renew SB login}
\usage{
sb_renew_login(username, password, verbose = FALSE)
}
\arguments{
\item{username}{Your ScienceBase/myUSGS username, usually an email address}

\item{password}{Your ScienceBase/myUSGS password}
}
\description{
Uses username and password to log in if session is expired or was never
authorized. Otherwise, pings SB to keep the session fresh.
}
\keyword{internal}

