% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_sb.R
\name{sb_require_login}
\alias{sb_require_login}
\title{Throw error or warning if SB login is inactive}
\usage{
sb_require_login(on_nonauth = c("stop", "warn", "continue", "login"),
  verbose = FALSE)
}
\arguments{
\item{on_nonauth}{character giving the action to take (stop or warn) if the 
session is not currently authorized}
}
\description{
Default is an error
}
\keyword{internal}

