% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_dataframe.R
\name{search_dataframe}
\alias{search_dataframe}
\title{Search within a data.frame for text matches to the constraints}
\usage{
search_dataframe(dataframe, constraints, match_case, fixed)
}
\arguments{
\item{constraints}{named vector of the constraints, e.g., 
c(site_name='nwis_0806'), where the names are column names in data.frame. 
constraints with NA values will be excluded}

\item{match_case}{logical. should the case be matched exactly?}

\item{fixed}{logical. As in \code{\link{grepl}}, TRUE to require an exact
match and FALSE to use regular expressions}
}
\value{
vector of T/F values where T indicates that the row of the dataframe 
  is matched.
}
\description{
Wrapper for grepl
}
\keyword{internal}
