% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metab_models.R
\name{search_metab_models}
\alias{search_metab_models}
\title{Search the metab_models}
\usage{
search_metab_models(model_name = NA, title = NA, row = NA, site = NA,
  date = NA, tag = NA, strategy = NA,
  metab_models = list_metab_models(), match_case = FALSE, fixed = TRUE)
}
\arguments{
\item{model_name}{text to search for within the entire model_name}

\item{title}{text to search for within the title portion of the model_name}

\item{row}{text to search for within the row portion of the model_name}

\item{site}{text to search for within the site portion of the model_name}

\item{date}{text to search for within the date portion of the model_name}

\item{tag}{text to search for within the tag portion of the model_name}

\item{strategy}{text to search for within the strategy portion of the
model_name}

\item{metab_models}{vector of model_names to search. The default is to pull
this list from ScienceBase}

\item{match_case}{logical. should the case be matched exactly?}

\item{fixed}{logical. As in \code{\link{grepl}}, TRUE to require an exact
match and FALSE to use regular expressions}
}
\description{
Search the metab_models
}
\examples{
\dontrun{
search_metab_models(site='nwis_08062500')
search_metab_models('indy', row="68.", fixed=FALSE)
search_metab_models(strategy='NIGHT', site='411955088280601')
}
}

