% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_metab_runs.R
\name{search_metab_runs}
\alias{search_metab_runs}
\title{Search the metab_runs}
\usage{
search_metab_runs(metab_run_title = NA, date = NA, tag = NA,
  strategy = NA, metab_runs = list_metab_runs(), match_case = FALSE,
  fixed = TRUE)
}
\arguments{
\item{metab_run_title}{text to search for within the entire metab_run_title}

\item{date}{text to search for within the date portion of the model_name}

\item{tag}{text to search for within the tag portion of the model_name}

\item{strategy}{text to search for within the strategy portion of the 
model_name}

\item{metab_runs}{vector of metab_run_titles to search. The default is to
pull this list from ScienceBase}

\item{match_case}{logical. should the case be matched exactly?}

\item{fixed}{logical. As in \code{\link{grepl}}, TRUE to require an exact
match and FALSE to use regular expressions}
}
\description{
Search the metab_runs
}
\examples{
\dontrun{
search_metab_runs(strategy='PR_FIXED')
search_metab_runs(date='^1512', fixed=FALSE)
search_metab_runs(tag='(0)\\\\.(0)\\\\.(1)[2-4]', fixed=FALSE)
}
}
