% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_var_src_codes.R
\name{search_var_src_codes}
\alias{search_var_src_codes}
\title{Search the var_src_codes}
\usage{
search_var_src_codes(var = NA, src = NA, var_src = NA, units = NA,
  var_descrip = NA, src_descrip = NA, data_type = NA, p_code = NA,
  metab_var = NA, priority = NA, var_src_codes = get_var_src_codes(),
  match_case = FALSE, fixed = TRUE)
}
\arguments{
\item{var}{text to search for in the var column}

\item{src}{text to search for in the src column}

\item{var_src}{text to search for in the var_src column}

\item{units}{text to search for in the units column}

\item{var_descrip}{text to search for in the var_descrip column}

\item{src_descrip}{text to search for in the src_descrip column}

\item{data_type}{text to search for in the data_type column}

\item{p_code}{text to search for in the p_code column}

\item{metab_var}{text to search for in the metab_var column}

\item{priority}{text to search for in the priority column}

\item{var_src_codes}{table of var_src_codes to search}

\item{match_case}{logical. should the case be matched exactly?}

\item{fixed}{logical. As in \code{\link{grepl}}, TRUE to require an exact
match and FALSE to use regular expressions}
}
\description{
Search the var_src_codes
}
\examples{
search_var_src_codes(units='mgO2 L^-1')
}

