% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_scheme.R
\name{set_scheme}
\alias{set_scheme}
\title{Set the scheme name of the project}
\usage{
set_scheme(scheme)
}
\arguments{
\item{scheme}{character name of the ScienceBase project}
}
\description{
Set the scheme name of the project
}
