% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_nwis_df.R
\name{split_site}
\alias{split_site}
\title{splot site name into siteID (used for NWIS site numbers)}
\usage{
split_site(site)
}
\arguments{
\item{site}{a valid powstreams site (see \code{\link{get_sites}})}
}
\value{
a site identifier. Splits on "-"
}
\description{
splot site name into siteID (used for NWIS site numbers)
}

