% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{stage_calc_ts}
\alias{stage_calc_ts}
\title{stage calculated or modeled data into a time series file}
\usage{
stage_calc_ts(sites, var, src, folder = tempdir(), verbose = FALSE, ...)
}
\arguments{
\item{sites}{a character vector of valid NWIS site IDs}

\item{var}{short name of variable as in \code{get_var_codes(out='var')}}

\item{src}{short name of src as in \code{get_src_codes()}}

\item{folder}{a folder to place the file outputs in (defaults to temp directory)}

\item{verbose}{provide verbose output (currently not implemented)}

\item{...}{additional arguments passed to \code{\link[geoknife]{geoknife}} and \code{\link[unitted]{write_unitted}}}
}
\value{
a file handle for time series file created
}
\description{
accept calculated/modeled data as a data.frame and return a file handle
}
\examples{
\dontrun{
files <- stage_calc_ts(sites="nwis_08062500", var="suntime", src="calcLon", verbose=TRUE)
str(read_ts(files))
files <- stage_calc_ts(sites="nwis_01484680", var="dosat", src="calcGG", verbose=TRUE)
}
}

