% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_nldas_ts.R
\name{stage_gldas_ts}
\alias{stage_gldas_ts}
\title{stage gldas data into a time series file}
\usage{
stage_gldas_ts(sites, var, times, folder = tempdir(), version = c("rds",
  "tsv"), verbose = FALSE, url, ...)
}
\arguments{
\item{sites}{a character vector of valid NWIS site IDs}

\item{var}{short name of var as in
\code{unique(get_var_src_codes(out='var'))}}

\item{times}{a length 2 vector of POSIXct dates}

\item{folder}{a folder to place the file outputs in (defaults to temp 
directory)}

\item{version}{character string indicating whether you want to stage the 
\code{ts} as a .tsv or .rds}

\item{verbose}{logical. provide verbose output?}

\item{url}{for web dataset. If missing, uses geoknife's "gldas" url}

\item{...}{additional arguments passed to \code{\link[geoknife]{geoknife}}}
}
\value{
a file handle for time series file created
}
\description{
get data from nldas and return a file handle
}
\examples{
\dontrun{
files <- stage_gldas_ts(sites = c("nwis_06893820","nwis_01484680"), var = "baro", 
  times = c('2014-01-01 00:00','2014-01-01 05:00'), verbose = TRUE)
read_ts(files[1])
}
}

