% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_indy_site.R
\name{stage_indy_site}
\alias{stage_indy_site}
\title{Generate a list of files to post as a custom ('indy') site}
\usage{
stage_indy_site(site_num, data, info = as.character(NA),
  long_name = as.character(NA), lat = u(as.numeric(NA), "degN"),
  lon = u(as.numeric(NA), "degE"), alt = u(as.numeric(NA), "ft"),
  var_format = c("mda.streams", "streamMetabolizer"), collapse_const = TRUE,
  folder = tempdir())
}
\arguments{
\item{site_num}{The id, not necessarily numeric that is to follow 'indy_' in 
the site_name}

\item{data}{data.frame of data to parse and post}

\item{info}{character string describing the primary source of data for this 
site}

\item{long_name}{Longer character string describing the site, or blank if 
site_name is prefixed with 'nwis' (in which case this will be looked up)}

\item{lat}{latitude in degN}

\item{lon}{longitude in degE}

\item{alt}{altitude in ft}

\item{var_format}{What format are the data column names in - mda.streams 
format (e.g., sitetime, doobs, dosat) or streamMetabolizer format (e.g., 
local.time. DO.obs, DO.sat)?}

\item{collapse_const}{If a data column has all identical values, should these
be collapsed to a single const row in the ts file?}

\item{folder}{the folder in which to save the staged information}
}
\description{
This function is designed for truly isolated (non-NWIS, etc.) sites. The site
that is created will be named 'indy_[site_num]', where 'indy' is a recognized
prefix in mda.streams. If an NWIS gage is available, even if some data is not
from NWIS, the site should be established as an 'nwis' site (see post_site) 
as for other NWIS sites; this will enable lookups of other NWIS data and 
metadata for the site.
}

