% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_ldas_ts.R
\name{stage_ldas_ts}
\alias{stage_ldas_ts}
\title{stage gldas data into a time series file}
\usage{
stage_ldas_ts(sites, var, p_code, times, folder, version, verbose, url,
  expected_units, ...)
}
\arguments{
\item{sites}{a character vector of valid NWIS site IDs}

\item{var}{short name of var as in
\code{unique(get_var_src_codes(out='var'))}}

\item{times}{a length 2 vector of POSIXct dates}

\item{folder}{a folder to place the file outputs in (defaults to temp 
directory)}

\item{version}{character string indicating whether you want to stage the 
\code{ts} as a .tsv or .rds}

\item{verbose}{logical. provide verbose output?}

\item{url}{for web dataset. If missing, uses geoknife's "gldas" url}

\item{...}{additional arguments passed to \code{\link[geoknife]{geoknife}}}
}
\value{
a file handle for time series file created
}
\description{
get data from nldas and return a file handle
}
\keyword{internal}

