% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_meta.R
\name{stage_meta}
\alias{stage_meta}
\title{Create/modify sites file}
\usage{
stage_meta(sites = list_sites(), folder = tempdir(), verbose = FALSE)
}
\arguments{
\item{sites}{a vector of sites to acquire/update metadata for}

\item{folder}{the folder in which to save the metadata file}

\item{verbose}{logical. print status messages?}
}
\description{
Create/modify sites file
}

