% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_basic.R
\name{stage_meta_basic_indy}
\alias{stage_meta_basic_indy}
\title{Get data for Indy (custom/independently collected data) sites}
\usage{
stage_meta_basic_indy(sites_meta, empty_meta, verbose = FALSE)
}
\arguments{
\item{sites_meta}{a data.frame of site_names and parsed site_names, as in the
opening lines of stage_meta_basic}

\item{empty_meta}{a data.frame showing how an empty result should look}

\item{verbose}{logical. give status messages?}
}
\description{
Helper to stage_meta_basic. Copies data from meta_indy into meta_basic where 
this is available.
}
