% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_basic.R
\name{stage_meta_basic_nwis}
\alias{stage_meta_basic_nwis}
\title{Get data for NWIS sites}
\usage{
stage_meta_basic_nwis(sites_meta, empty_meta, verbose = FALSE)
}
\arguments{
\item{sites_meta}{a data.frame of site_names and parse site_names, as in the
opening lines of stage_meta_basic}

\item{empty_meta}{a data.frame showing how an empty result should look}

\item{verbose}{logical. give status messages?}
}
\description{
Helper to stage_meta_basic. Download/clean up columns of useful data for NWIS
sites
}
\examples{
\dontrun{
sites <- list_sites()[655:665]
sites_meta <- tibble::tibble(
  site_name=sites, 
  site_database=parse_site_name(sites, out='database'),
  site_num=parse_site_name(sites, out='sitenum')) \%>\%
  as.data.frame() \%>\% # unitted_data.frames work better than unitted_tbl_dfs for now
  u()
str(mda.streams:::stage_meta_basic_nwis(sites_meta, verbose=TRUE))
}
}
\keyword{internal}

