% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_dvqcoefs.R
\name{stage_meta_dvqcoefs}
\alias{stage_meta_dvqcoefs}
\title{Create/modify file of depth v discharge scaling relationships}
\usage{
stage_meta_dvqcoefs(folder = tempdir(), verbose = FALSE)
}
\arguments{
\item{folder}{the folder in which to save the metadata file}

\item{verbose}{logical. print status messages?}
}
\description{
Pull in scaling coefficients from Jud Harvey
}
