% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_indy.R
\name{stage_meta_indy}
\alias{stage_meta_indy}
\title{Create/modify the meta_indy file}
\usage{
stage_meta_indy(rows = u(data.frame(site_name = "indy_xxxx", long_name =
  as.character(NA), lat = u(as.numeric(NA), "degN"), lon = u(as.numeric(NA),
  "degE"), alt = u(as.numeric(NA), "ft"), info = as.character(NA))),
  on_exists = c("stop", "replace"), folder = tempdir(), verbose = FALSE)
}
\arguments{
\item{rows}{a data frame with one or more rows, having the same columns as 
the existing meta_indy (if there is one). These rows will be inserted 
according to their site_name values. Columns with names that overlap with 
other metadata tables, as prefixed by the table name (e.g., 'basic.lat'), 
will be copied to those tables when the corresponding stage_meta_... 
function is called (except that this copying is currently only implemented 
for the 'basic' metadata table).}

\item{on_exists}{character indicating how to handle new rows whose site_name 
is already present in the current meta_ file If 'stop', an error will be 
thrown if even one value is duplicated. If 'replace', a message will be 
given and existing rows with site_name in rows$site_name will be replaced
by the new rows.}

\item{folder}{where to store the staged file}

\item{verbose}{logical. give status messages?}
}
\description{
Create/modify the meta_indy file
}
