% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_manual.R
\name{stage_meta_manual}
\alias{stage_meta_manual}
\title{Create/modify file of manually-determined site information}
\usage{
stage_meta_manual(folder = tempdir(), verbose = FALSE)
}
\arguments{
\item{folder}{the folder in which to save the metadata file}

\item{verbose}{logical. print status messages?}
}
\description{
Pull in user review data from Google Spreadsheet at
}

