% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_meta_metabinput.R
\name{stage_meta_metabinput}
\alias{stage_meta_metabinput}
\title{Create/modify the meta_metabinput file}
\usage{
stage_meta_metabinput(model = "metab_mle", model_args = "list()",
  folder = tempdir(), verbose = FALSE)
}
\arguments{
\item{model}{the model from which to determine the input dataset column needs}

\item{model_args}{the model args to use, if applicable, in deriving the input dataset}

\item{folder}{where to store the staged file}

\item{verbose}{logical. give status messages?}
}
\description{
Create/modify the meta_metabinput file
}

