% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_meta_nawqahst.R
\name{stage_meta_nawqahst}
\alias{stage_meta_nawqahst}
\title{Acquire data from the Accumulated and Reach Catchment Land Variables dataset}
\usage{
stage_meta_nawqahst(types = c("AC_RUNOFF", "Ac_popd10", "AC_NLCD11"),
  folder = "temp")
}
\arguments{
\item{types}{the metadata types to be acquired. This type will be identical
(even in case) to the name of the zip file from which the data will be
acquired. E.g., AC_NLCD11 is the type for
https://www.sciencebase.gov/catalog/item/534c35d0e4b0af6611b1d0d0,
Accumulated areas of NLCD 2011 Land-Use Classes}

\item{folder}{the folder in which to store both the full table from Mike
Wieczorek's site and the formatted, subsetted table destined for our SB
directory}
}
\description{
Mike Wieczorek's database, Accumulated and Reach Catchment Land Variables
(abbreviated ARCLV here), contains data computed for every NHDPlus reach,
identifiable by ComID. This function pulls data from one ARCLV table, subsets
to just the reaches relevant to the Powell Center study, and prepares that
table to be posted as a metadata table in the Powell Center SB directory.
}

