% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_nawqahst_temp.R
\name{stage_meta_nawqahst_temp}
\alias{stage_meta_nawqahst_temp}
\title{Import data that Ted has acquired from the Accumulated and Reach Catchment 
Land Variables dataset}
\usage{
stage_meta_nawqahst_temp(types = c("climate", "hydrol", "soil", "landcover"),
  folder = tempdir())
}
\arguments{
\item{types}{the metadata table to be acquired}

\item{folder}{the folder in which to store the formatted table destined for
our SB directory}
}
\description{
This is a temporary step toward the long-term goal of having every data 
import step fully scripted. In the meantime, see 
stream_metab_usa/starter_files/data_scripts for what Ted emailed to Alison in
July 2015.
}

