% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_meta.R
\name{stage_meta_nwis}
\alias{stage_meta_nwis}
\title{Get data for NWIS sites}
\usage{
stage_meta_nwis(sites_meta, verbose = FALSE)
}
\arguments{
\item{sites_meta}{a data.frame of site_names and parse site_names, as in the
opening lines of stage_meta}

\item{verbose}{logical. give status messages?}
}
\description{
Helper to stage_meta. Download/clean up columns of useful data for NWIS
sites
}
\keyword{internal}

