% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_mfoot.R
\name{stage_meta_struct}
\alias{stage_meta_struct}
\title{Create/modify file of interferences within the metabolic footprints}
\usage{

  stage_meta_struct(struct_file = "../stream_metab_usa/1_spatial/in/CONF_struct_170407.csv",
  folder = tempdir(), verbose = FALSE)
}
\arguments{
\item{struct_file}{path to the file containing data on distances to
structural interferences - NPDES, canals, and dams. Description of that
file: "A -9999 means that there weren't any of that feature in the basin.
i.e. no dam or no NPDES etc.  Other numbers express the straight-line
geodesic distance between the probe and the feature in meters"}

\item{folder}{the folder in which to save the metadata file}

\item{verbose}{logical. print status messages?}
}
\description{
Create/modify file of interferences within quantiles of the metabolic
footprint
}
