% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_meta.R
\name{stage_meta_styx}
\alias{stage_meta_styx}
\title{Get data for Styx (simulated data) sites}
\usage{
stage_meta_styx(sites_meta, verbose = FALSE)
}
\arguments{
\item{sites_meta}{a data.frame of site_names and parsed site_names, as in the
opening lines of stage_meta}

\item{verbose}{logical. give status messages?}
}
\description{
Helper to stage_meta. Create dummy columns of metadata for Styx sites.
}
\keyword{internal}

