% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_metab_model.R
\name{stage_metab_model}
\alias{stage_metab_model}
\title{Stage metabolism model outputs into individual metab_model .RData files}
\usage{
stage_metab_model(title, metab_outs, folder = tempdir(),
  version = c("original", "modern"), verbose = FALSE)
}
\arguments{
\item{title}{the title of the metab modeling run whose results are in
metab_outs}

\item{metab_outs}{a single metab_model or a list of metab_models from which
to extract the metab_model objects}

\item{folder}{a folder to place the file outputs in (defaults to temp
directory)}

\item{version}{character indicating whether this is the original verison of
this metab_model or a modernized one}

\item{verbose}{provide verbose output (currently not implemented)}
}
\value{
a character vector of file handles
}
\description{
Parse a list of metab_outs into individual .RData files
}

