% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_metab_ts.R
\name{stage_metab_ts}
\alias{stage_metab_ts}
\title{Stage metabolism model outputs into ts files}
\usage{
stage_metab_ts(metab_outs, vars = c("gpp", "er", "K600"),
  folder = tempdir(), verbose = FALSE)
}
\arguments{
\item{metab_outs}{a single metab_model or a list of metab_models from which
to extract GPP, ER, and K600}

\item{vars}{one or more strings in \code{c("gpp","er","K600")} naming the
variables to extract and post}

\item{folder}{a folder to place the file outputs in (defaults to temp 
directory)}

\item{verbose}{provide verbose output (currently not implemented)}
}
\value{
a character vector of file handles
}
\description{
Parse metabolism model outputs into gpp, er, and K600
}
