% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_nldas_ts.R
\name{stage_nldas_ts}
\alias{stage_nldas_ts}
\title{stage nldas data into a time series file}
\usage{
stage_nldas_ts(sites, var, times, folder = tempdir(), verbose = FALSE, ...)
}
\arguments{
\item{sites}{a character vector of valid NWIS site IDs}

\item{var}{short name of var as in
\code{\link{get_var_codes(out='var')}}}

\item{times}{a length 2 vector of POSIXct dates}

\item{folder}{a folder to place the file outputs in (defaults to temp
directory)}

\item{verbose}{logical. provide verbose output?}

\item{...}{additional arguments passed to \code{\link[geoknife]{geoknife}}
and \code{\link[unitted]{write_unitted}}}
}
\value{
a file handle for time series file created
}
\description{
get data from nldas and return a file handle
}
\examples{
\dontrun{
files <- stage_nldas_ts(sites = c("nwis_06893820","nwis_01484680"), var = "baro",
  times = c('2014-01-01 00:00','2014-01-01 05:00'), verbose = TRUE)
read_ts(files[1])
}
}

