% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_nwis_sitelist.R
\name{stage_nwis_sitelist}
\alias{stage_nwis_sitelist}
\title{generate nwis site_ids for given p_codes}
\usage{
stage_nwis_sitelist(p_codes, stateCd, folder = NULL)
}
\arguments{
\item{p_codes}{a character vector of NWIS p_codes}

\item{stateCd}{character vector of state codes, or "all" for all data.}

\item{folder}{directory path, or NULL, indicating where to save the file or
(NULL) to return it as a character vector}
}
\value{
a character vector of NWIS sites, appended with 'nwis_'
}
\description{
finds all NWIS sites that meet given data requirements
}
\examples{
\dontrun{
sites <- stage_nwis_sitelist(p_codes = c('00095', '00060', '00010', '00300'),
                             stateCd=c("wi","mi"))
}
}

