% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize_metab_inputs.R
\name{summarize_metab_inputs}
\alias{summarize_metab_inputs}
\title{Summarize a timeseries}
\usage{
summarize_metab_inputs(config, rows, out = c("site", "model", "tag",
  "strategy", "date", "start_date", "end_date", "true_start", "true_end",
  "num_dates", "num_rows", "num_complete", "modal_timestep",
  "num_modal_timesteps"), verbose = TRUE)
}
\arguments{
\item{config}{a config file/data.frame from which datasets should be
constructed for summarization.}

\item{rows}{missing, integer, or vector of integers. The row number[s] of the
config data.frame to use for this particular data summary.}

\item{out}{a list of one or more outputs to include in the summary dataframe}

\item{verbose}{logical. supply status messages?}
}
\description{
Summarize a timeseries
}

