% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_metab_model.R
\name{summarize_metab_model}
\alias{summarize_metab_model}
\title{Summarize a single metabolism model}
\usage{
summarize_metab_model(models, on_local_exists = "replace",
  out = c("num_dates", "num_complete", "KvR_kendall_cor", "GPP_pct_neg",
  "ER_pct_pos", "K600_pct_neg", "GPP_q0", "GPP_q10", "GPP_q50", "GPP_q90",
  "GPP_q100", "ER_q0", "ER_q10", "ER_q50", "ER_q90", "ER_q100", "K600_q0",
  "K600_q10", "K600_q50", "K600_q90", "K600_q100", "DO_RMSE_q0", "DO_RMSE_q10",
  "DO_RMSE_q50", "DO_RMSE_q90", "DO_RMSE_q100"))
}
\arguments{
\item{models}{either a character vector containing model_names or a named 
list of metab model objects where the list element names are model_names}

\item{on_local_exists}{character indicating what to do if the folder already 
contains a file with the intended download name}

\item{out}{a list of one or more outputs to include in the summary dataframe,
in addition to the model_name and parsed name columns.}
}
\description{
Returns a data.frame of metrics, one row per model_name
}
\examples{
\dontrun{
login_sb()
mymms <- c("nwis_11462500-4-150729 0.0.5 bob_sites_test_mle",
  "nwis_11462500-349-150730 0.0.6 nighttime_k_plus_data",
  "nwis_11462500-349-150730 0.0.7 MLE_for_PRK_wHarvey_and_sw")
summarize_metab_model(mymms) # be patient
}
}
