% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summarize_ts.R
\name{summarize_ts}
\alias{summarize_ts}
\title{Summarize a timeseries}
\usage{
summarize_ts(var_src, site_name, ...)
}
\arguments{
\item{var_src}{character vector of dataset types to summarize.}

\item{site_name}{character vector of the same length as var_src specifying the
site or sites (e.g., "nwis_24309857") where each timeseries dataset should
be sought.}

\item{...}{other args passed to \code{\link[sbtools]{session_check_reauth}}}
}
\description{
Summarize a timeseries
}
\examples{
\dontrun{
summarize_ts("doobs_nwis", "nwis_01021050")
}
}

