% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ts_files.R
\name{summarize_ts_files}
\alias{summarize_ts_files}
\title{Get a table of information on ts files for the specified var_src (or all 
var_srces)}
\usage{
summarize_ts_files(var_src)
}
\arguments{
\item{var_src}{one or more var_srces to include in the query. if missing, all
var_src values from get_var_src_codes(out='var_src') will be included (may 
be slow!)}
}
\description{
The uploaded_by column is only available when you are logged in to
ScienceBase.
}

