% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_calc_ts.R
\name{try_calc_ts}
\alias{try_calc_ts}
\title{Attempt to stage and post a calc_ts for all available sites}
\usage{
try_calc_ts(var_src, sb_user, sb_pass, done_after = Sys.time() -
  as.difftime(3, units = "days"), done_version = "rds")
}
\arguments{
\item{var_src}{the var_src to stage+post}

\item{sb_user}{the sciencebase username}

\item{sb_pass}{the sciencebase password}

\item{done_after}{posixct datetime - sites will not be considered 'done' 
unless there is a timeseries file available from after this datetime}

\item{done_version}{character in c('rds','tsv') - sites will not be
considered 'done' unless the timeseries file already constrained by
done_after is also one of the specified version}
}
\description{
Attempt to stage and post a calc_ts for all available sites
}
