% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verify_config.R
\name{verify_config}
\alias{verify_config}
\title{Verify that a config file row has the requisite info}
\usage{
verify_config(config)
}
\arguments{
\item{config}{a config df to verify}
}
\value{
TRUE if is valid, FALSE if not
}
\description{
Checks the config row for valid src-site-logic pairs with respect to a given
model
}

