% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verify_ts.R
\name{verify_ts}
\alias{verify_ts}
\title{verify timeseries data.frame format and contents}
\usage{
verify_ts(data, variable, checks = c("cols", "tz", "units", "variable"))
}
\arguments{
\item{data}{a \code{\link[unitted]{unitted}} data.frame with date and value}

\item{variable}{optional name for timeseries. If supplied, will check variable column name}

\item{checks}{tests to run on the data ('cols', 'tz', 'units', 'variable')}
}
\value{
TRUE if is valid, FALSE if not
}
\description{
Checks to see a given data.frame meets project criteria
}
\examples{
\dontrun{
files <- stage_nwis_ts(sites = c("nwis_06893820","nwis_01484680"), variable = "doobs",
                 times = c('2014-01-01','2014-02-01'))
ts <- read_ts(files[1])
verify_ts(ts)
verify_ts(ts, variable = 'doobs', checks = c('cols', 'units', 'variable'))
}
}
\keyword{internal}

