% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{verify_var_src}
\alias{verify_var_src}
\title{Check whether a var_src is listed in get_var_src_codes()}
\usage{
verify_var_src(var_var_src, src, on_fail = warning)
}
\arguments{
\item{var_var_src}{may be either a variable, e.g. "doobs", or a var_src, e.g.
"doobs_nwis".}

\item{src}{a data source; optional if var_var_src is specified with both
variable and source}

\item{on_fail}{the function to apply to an information message (character) if
the var_src is not valid}
}
\value{
a single TRUE if successful, else on_fail(msg) followed by a vector
  of T/F values with F for invalid var_srces
}
\description{
Check whether a var_src is listed in get_var_src_codes()
}

