% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_file.R
\name{view_folder}
\alias{view_folder}
\title{Open a folder or a file's folder}
\usage{
view_folder(path)
}
\arguments{
\item{path}{a file path}
}
\description{
Opens the specified folder, or the containing folder of the specified file,
in the file explorer
}
