% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_config.R
\name{write_config}
\alias{write_config}
\title{Write a config file}
\usage{
write_config(config, file)
}
\arguments{
\item{config}{a config data.frame to write}

\item{file}{the file name/path to write to}
}
\description{
Write a config file
}
