% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts}
\alias{write_ts}
\title{write timeseries format compressed file}
\usage{
write_ts(data, site, var, src, folder)
}
\arguments{
\item{data}{unitted data.frame}

\item{site}{the full site name, e.g., 'nwis_06893820'}

\item{var}{the variable name of the output, e.g., 'doobs'}

\item{src}{the source of the data, e.g., 'nwis' or 'nldas'}

\item{folder}{the folder to write the file in}
}
\description{
write timeseries format compressed file
}
\keyword{internal}

