% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_calc_ts_needs.R
\name{build_calc_ts_needs}
\alias{build_calc_ts_needs}
\title{Build and return the table calc_ts_needs}
\usage{
build_calc_ts_needs(var, src)
}
\description{
Build and return the table calc_ts_needs
}

