% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{calc_ts_par_calcLatSw}
\alias{calc_ts_par_calcLatSw}
\title{Internal - calculate par by merging modeled (calcLat) and observed (calcSw) data}
\usage{
calc_ts_par_calcLatSw(utctime, parlat, parsw, latitude, longitude)
}
\arguments{
\item{utctime}{the DateTime with tz of UTC}

\item{parlat}{PAR in umol m^2 s^-1}
}
\description{
Internal - calculate par by merging modeled (calcLat) and observed (calcSw) data
}
\keyword{internal}

