% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config_to_data.R
\name{config_preds_to_data_column}
\alias{config_preds_to_data_column}
\title{Get predictions of metabolism or DO from a model}
\usage{
config_preds_to_data_column(var, site, model_src = src, src_type = "SB")
}
\arguments{
\item{var}{the variable name to retrieve}

\item{model_src}{the model_name or model file name}

\item{src_type}{character in c("SB","file") indicating whether model_src is a
model on sciencebase or a local file name}
}
\description{
Most of this code is the same whether it's a pred (the name of a model on
sciencebase) or a pred_file (the name of a local file name)
}
\keyword{internal}

