% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config_to_data.R
\name{config_to_data}
\alias{config_to_data}
\title{COnvert config specifications into a data input table}
\usage{
config_to_data(config_row, row_num, metab_fun, metab_args,
  on_error = c("stop", "warn", "quiet"))
}
\arguments{
\item{config_row}{a 1-row config data.frame}

\item{row_num}{the row number/name that this config_row held in the original
config}

\item{metab_fun}{a metabolism modeling function}

\item{metab_args}{a list of arguments (possibly just \code{list()}) that will
later be passed to the metab_fun}

\item{on_error}{character. What to do if data cannot be acquired or merged?
If 'stop' or 'warn', the appropriate condition is thrown. If 'warn' or
'quiet', the function [also] returns a character of errors and/or a
character of warnings, attached as attributes of the output object}
}
\value{
NA or a unitted data.frame, possibly with attributes "errors" and/or
  "warnings" attached (see on_error)
}
\description{
Turns a single config row into a data.frame of input data for the specified
metabolism modeling function
}
\examples{
\dontrun{
depth_file <- download_ts("depth_calcDischHarvey", "nwis_04087142",
  on_local_exists="replace")
config <- stage_metab_config(tag="0.0.1", strategy="try stage_metab_config",
  site="nwis_04087142", dosat=choose_data_source("dosat", "nwis_04087142",
    logic="simple dosat", type="const", src="12,mg L^-1"),
  depth=choose_data_source("depth", "nwis_04087142", logic="local file",
    type="file", src=depth_file), filename=NULL)
cdat <- config_to_data(config[1,], row_num=1, metab_fun=metab_mle, metab_args=list())
names(cdat)
head(cdat[['data']])

login_sb()
site="nwis_01646000"
config <- stage_metab_config(tag="0.0.1", strategy="try stage_metab_config",
  model="metab_Kvpred", site=site, filename=NULL,
  sitetime=choose_data_source("sitetime", site, logic="unused var"),
  doobs=choose_data_source("doobs", site, logic="unused var"),
  dosat=choose_data_source("dosat", site, logic="unused var"),
  depth=choose_data_source("depth", site, logic="unused var"),
  wtr=choose_data_source("wtr", site, logic="unused var"),
  par=choose_data_source("par", site, logic="unused var"),
  sitedate=choose_data_source("sitedate", site, logic="priority local"),
  K600=choose_data_source("K600", site, logic="nighttime", src="0.0.6", type="pred"),
  dischdaily=choose_data_source("dischdaily", site, logic="priority local"),
  velocdaily=choose_data_source("velocdaily", site, logic="priority local"),
  omit_incomplete=FALSE)
cdat <- config_to_data(config[1,], row_num=1,
  metab_fun=streamMetabolizer::metab_Kvpred, metab_args=list())
}
}

