% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config_to_metab.R
\name{config_to_metab}
\alias{config_to_metab}
\title{Actually run the model as specified by the configuration arguments}
\usage{
config_to_metab(config, rows)
}
\arguments{
\item{config}{data.frame, or file/filename to read a data.frame from. As an
alternative to all preceding arguments, this single config argument may be
passed in. The data.frame columns must correspond precisely to the list of
preceding arguments.}

\item{rows}{missing, integer, or vector of integers. The row number[s] of the
config data.frame to use for this particular model run.}
}
\description{
Actually run the model as specified by the configuration arguments
}
\examples{
\dontrun{
config_to_metab(config=stage_metab_config(
  tag="0.0.1", strategy="try stage_metab_config",
  site="nwis_04087142", filename=NULL))
}
}

