% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_site.R
\name{create_site}
\alias{create_site}
\title{Create new site in SB}
\usage{
create_site(site, replace_existing = FALSE, ...)
}
\arguments{
\item{site}{Local site ID (e.g. nwis_09238475)}

\item{replace_existing}{logical. Should an item that already exists be
replaced?}

\item{...}{Additional parameters supplied to \code{\link[sbtools]{session_check_reauth}}}
}
\value{
an item list
}
\description{
Create new site in SB
}
\examples{
\dontrun{
locate_site("nwis_08437710", format="url")
create_site("nwis_08437710") # returns list()
create_site("nwis_08437710", replace_existing=TRUE)
}
}
\author{
Corinna Gries
}

