% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_site.R
\name{create_site}
\alias{create_site}
\title{Create new site in SB}
\usage{
create_site(site, session, replace_existing = FALSE)
}
\arguments{
\item{site}{Local site ID (e.g. nwis_09238475)}

\item{session}{Session object from \code{\link[sbtools]{authenticate_sb}}}

\item{skip_exists}{boolean for skip creation if item already exists.
if TRUE, site will not be created and an empty list will be returned}
}
\value{
and item list
}
\description{
Create new site in SB
}
\author{
Corinna Gries
}

