% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_site.R
\name{create_site}
\alias{create_site}
\title{Create new site in SB}
\usage{
create_site(site, session)
}
\arguments{
\item{site}{Local site ID (e.g. NWIS_09238475)}

\item{session}{Session object from \link{authenticate_sb}}
}
\description{
Create new site in SB
}
\author{
Corinna Gries
}

