% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_meta.R
\name{delete_meta}
\alias{delete_meta}
\title{Delete a metadata item-file combination}
\usage{
delete_meta(types, files_only = FALSE, verbose = TRUE)
}
\arguments{
\item{types}{character. One or more types of meta file}

\item{files_only}{logical. Should only the files, not the item, be delted?}

\item{verbose}{logical. Should status messages be given?}
}
\description{
Deletes a metadata object
}
\keyword{internal}
