% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_site.R
\name{delete_site}
\alias{delete_site}
\title{Delete a site and all its files.}
\usage{
delete_site(sites, children_only = FALSE, verbose = TRUE)
}
\arguments{
\item{sites}{one or more site IDs such as "nwis_08437710" to look up and 
delete}

\item{children_only}{logical. If TRUE, only the child items in the site 
folder will be deleted, leaving an empty site folder}

\item{verbose}{logical. Should status messages be given?}
}
\value{
anything passed back from the final item_rm if children_only==FALSE,
  or a pointer to the remaining item if children_only==TRUE
}
\description{
Delete a site and all its files.
}
\examples{
\dontrun{
login_sb()
set_scheme("mda_streams_dev")

sites <- c("nwis_00000000", "nwis_00000001", "nwis_00000002")
post_site(sites)
file_par2 <- stage_calc_ts(sites="nwis_00000000", var="par", src="simNew", verbose=TRUE,
  inputs=list(utctime=NA, value=unitted::u(rnorm(1, 10, 2), "umol m^-2 s^-1")))
post_ts(file_par2)
list_tses("nwis_00000000")
mda.streams:::delete_site(sites, children_only=TRUE)
locate_site(sites)
list_tses("nwis_00000000")
mda.streams:::delete_site(sites)
locate_site(sites)

set_scheme("mda_streams")
}
}
\keyword{internal}
