% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_site.R
\name{delete_site}
\alias{delete_site}
\title{Delete a site and all its files.}
\usage{
delete_site(sites, children_only = FALSE, verbose = TRUE)
}
\arguments{
\item{sites}{one or more site IDs such as "nwis_08437710" to look up and
delete}

\item{children_only}{logical. If TRUE, only the child items in the site
folder will be deleted, leaving an empty site folder}

\item{verbose}{logical. Should status messages be given?}
}
\value{
anything passed back from the final item_rm if children_only==FALSE,
  or a pointer to the remaining item if children_only==TRUE
}
\description{
Delete a site and all its files.
}
\examples{
\dontrun{
sbtools::authenticate_sb()
set_scheme("mda_streams_dev")

sites <- c("nwis_00000000", "nwis_00000001", "nwis_00000002")
post_site(sites)
mda.streams:::delete_site(sites)

set_scheme("mda_streams")
}
}
\keyword{internal}

