% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_site.R
\name{delete_site}
\alias{delete_site}
\title{Delete a site and all its files.}
\usage{
delete_site(sites, verbose = TRUE, ...)
}
\arguments{
\item{sites}{one or more site IDs such as "nwis_08437710" to look up and
delete}

\item{verbose}{logical. Should status messages be given?}

\item{...}{other args passed to \code{\link[sbtools]{session_check_reauth}}}
}
\value{
anything passed back from the final item_rm
}
\description{
Delete a site and all its files.
}
\examples{
\dontrun{
cat(locate_site("nwis_08437710", format="url"))
files <- sapply(list_datasets("nwis_08437710", type="ts"), function(var_src) {
  download_ts(var_src, "nwis_08437710")
})

# try delete_site
mda.streams:::delete_site("nwis_08437710")
mda.streams:::delete_site(c("nwis_08437710","nwis_03401385","nwis_02423397"))

# now repair the damage
post_site("nwis_08437710")
Sys.sleep(1)
post_ts(files[!is.na(files)])

# or if you REALLY messed it up:
files <- sapply(c("doobs","disch","stage","wtr"), function(var) {
  staged <- stage_nwis_ts("nwis_08437710", var,
    times=as.POSIXct(paste(c("2014-01-01","2015-01-01"), "00:00:00"), tz="UTC"))
  if(is.null(staged)) NA else staged
})
}
}
\keyword{internal}

