% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_ts.R
\name{delete_ts}
\alias{delete_ts}
\title{Delete a time series item and its data}
\usage{
delete_ts(variable, siteid, verbose = TRUE, ...)
}
\arguments{
\item{variable}{the variable name, e.g., "doobs"}

\item{siteid}{the site ID, e.g. "nwis_02322688", whose folder you want}

\item{verbose}{logical. Should status messages be given?}

\item{...}{other args passed to \code{\link[sbtools]{session_check_reauth}}}
}
\description{
Deletes timeseries objects specified by all combinations of variable and site
}
\examples{
\dontrun{
mda.streams:::delete_ts(c("doobs","wtr"), c("nwis_03271510", "nwis_412126095565201", "nwis_03409500"), verbose=TRUE)
}
}
\keyword{internal}

