% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_item_files.R
\name{download_item_files}
\alias{download_item_files}
\title{Download metadata data to local file destination}
\usage{
download_item_files(item_ids, item_names, files = NA, folder = tempdir(),
  on_remote_missing = c("stop", "return_NA"), on_local_exists = c("stop",
  "skip", "replace"))
}
\arguments{
\item{item_ids}{the item id[s]}

\item{item_names}{character names by which to refer to the respective items 
in messages}

\item{files}{list of character vector[s] of file names to download, one list 
element per item_ids. if the default is used, all files from each site will
be downloaded. may also be passed as a single character vector to be
replicated once for each item_id}

\item{folder}{string for a folder location where files should be saved}

\item{on_remote_missing}{character indicating what to do if the}

\item{on_local_exists}{character indicating what to do if the folder already 
contains a file with the intended download name}
}
\value{
file handle (character path) for the downloaded file, or NA if the 
  timeseries is unavailable on ScienceBase
}
\description{
Download a timeseries file to a user-specified (or temp file) location
}
\examples{
\dontrun{
download_item_files(item_id, on_local_exists="replace")
}
}
\author{
Alison P Appling, Corinna Gries, Jordan S Read, Luke A Winslow
}

