% Generated by roxygen2 (4.0.2): do not edit by hand
\name{download_timeseries}
\alias{download_timeseries}
\title{Download timeseries to local computer}
\usage{
download_timeseries(site, variable, destination = NULL, session = NULL)
}
\arguments{
\item{site}{A unique site ID eg [NWIS_392572094]}

\item{variable}{Timeseries identifiying string eg [ts_stage, ts_doobs, ts_wtr, etc]}

\item{destination}{Full folder path and file name}

\item{session}{Session object from \link{authenticate_sb}}
}
\description{

}
\author{
Corinna Gries
}

