% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download_ts.R
\name{download_ts}
\alias{download_ts}
\title{download timeseries data to local file destination}
\usage{
download_ts(site, variable, destination = NULL, session = NULL, ...)
}
\arguments{
\item{site}{a valid mda.streams site (see \link{get_sites})}

\item{variable}{a valid variable name for timeseries data (see \link{get_ts_variables})}

\item{destination}{string for a folder location}

\item{session}{a valid sciencebase session (see \code{\link[sbtools]{authenticate_sb}}).
Set \code{session = NULL} (default) for sites on sciencebase that are public.}

\item{...}{additional arguments passed to \code{\link[sbtools]{item_file_download}},
for example \code{overwrite_file}}
}
\value{
file handle for downloaded file
}
\description{
download a timeseries file to a user-specified (or temp file) location
}
\examples{
\dontrun{
download_ts(site = 'nwis_01018035', variable = 'doobs')
}
}
\author{
Corinna Gries, Jordan S Read, Luke A Winslow
}

