% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download_ts.R
\name{download_ts}
\alias{download_ts}
\title{download timeseries data to local file destination}
\usage{
download_ts(var_src, site_name, folder = tempdir(),
  on_remote_missing = c("stop", "return_NA"), on_local_exists = c("stop",
  "skip", "replace"), ...)
}
\arguments{
\item{var_src}{a valid variable name for timeseries data (see
\code{get_var_codes(out='var_src', type='ts')})}

\item{site_name}{a valid mda.streams site (see \link{get_sites})}

\item{folder}{string for a folder location}

\item{on_local_exists}{character indicating what to do if the folder already
contains a file with the intended download name}

\item{...}{additional arguments passed to
\code{\link[sbtools]{session_check_reauth}}}

\item{on_remote_exists}{character indicating what to do if the}
}
\value{
file handle (character path) for the downloaded file, or NA if the
  timeseries is unavailable on ScienceBase
}
\description{
download a timeseries file to a user-specified (or temp file)
  location
}
\examples{
\dontrun{
download_ts(var_src = 'doobs_nwis', site_name = 'nwis_06893300')
}
}
\author{
Corinna Gries, Jordan S Read, Luke A Winslow, Alison P Appling
}

