% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_item.R
\name{format_item}
\alias{format_item}
\title{Format an item as a ScienceBase ID or URL, as requested}
\usage{
format_item(item, format, browser)
}
\arguments{
\item{item}{a data.frame with a column for id (and optionally columns for
title, etc.) describing one or more ScienceBase items}

\item{format}{character indicating whether the folder should be returned as 
an ID or as a full URL}

\item{browser}{logical. Should the URL be opened in a browser?}
}
\description{
Internal helper to locate_item. Doesn't check the value of format - that's up
to the calling functions to do, for efficiency.
}
\keyword{internal}
