% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_meta.R
\name{get_meta}
\alias{get_meta}
\title{Get metadata from one or more meta_xxx files}
\usage{
get_meta(types = c("basic", "metabinput"), out = "all")
}
\arguments{
\item{types}{one or more metadata types to select and merge into a single
table.}

\item{out}{character vector or 'all'. if 'all', all columns from the selected
types will be returned. if anything else, the selected columns will be
returned.}
}
\value{
a metadata table
}
\description{
Keeps a locally cached copy to reduce the size and number of requests to SB.
If
}

