% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metab_data.R
\name{get_metab_data}
\alias{get_metab_data}
\title{Get a data.frame of metabolism input data}
\usage{
get_metab_data(site, sitetime = NA, doobs = NA, dosat = NA, depth = NA,
  wtr = NA, par = NA, disch = NULL, veloc = NULL, sitedate = NULL,
  doinit = NULL, gpp = NULL, er = NULL, K600 = NULL, K600lwr = NULL,
  K600upr = NULL, gppinit = NULL, erinit = NULL, K600init = NULL,
  dischdaily = NULL, velocdaily = NULL, start_date = NA, end_date = NA,
  tag = "0.0.0", strategy = "get_metab_data", date = Sys.time(),
  model = "metab_mle", model_args = "list()", simplify = TRUE,
  verbose = TRUE)
}
\arguments{
\item{site}{site names}

\item{sitetime}{Data Source for mean solar time. See Data Source Format 
below.}

\item{doobs}{Data Source for dissolved oxygen concentrations. See Data Source
Format below.}

\item{dosat}{Data Source for dissolved oxygen saturation concentrations. See 
Data Source Format below.}

\item{depth}{Data Source for mean stream depth. See Data Source Format below.}

\item{wtr}{Data Source for water temperature. See Data Source Format below.}

\item{par}{Data Source for light (photosynthetically available radiation, 
PAR). See Data Source Format below.}

\item{disch}{Data Source for unit-value stream discharge, for use in 
identifying daily priors or fixed values for K600. See Data Source Format 
below.}

\item{veloc}{Data Source for unit-value flow velocity, for use in identifying
daily priors or fixed values for K600. See Data Source Format below.}

\item{sitedate}{Data Source for the dates of interest. See Data Source Format
below.}

\item{doinit}{Data Source for the first DO observation on each date to model,
for use in data simulation. See Data Source Format below.}

\item{gpp}{Data Source for daily gross primary productivity rates for use in 
data simulation. See Data Source Format below.}

\item{er}{Data Source for ecosystem respiration rates for use in data 
simulation. See Data Source Format below.}

\item{K600}{Data Source for reaeration rates for use in data simulation. See 
Data Source Format below.}

\item{K600lwr}{Data Source for lower bound on reaeration rates for use in 
data simulation. See Data Source Format below.}

\item{K600upr}{Data Source for upper bound on reaeration rates for use in 
data simulation. See Data Source Format below.}

\item{gppinit}{Data Source for initial values of daily gross primary
productivity rates for use in likelihood maximization. See Data Source
Format below.}

\item{erinit}{Data Source for initial values of ecosystem respiration rates
for use in likelihood maximization. See Data Source Format below.}

\item{K600init}{Data Source for initial values of reaeration rates for use in
likelihood maximization. See Data Source Format below.}

\item{dischdaily}{Data Source for daily mean stream discharge, for use in 
identifying daily priors or fixed values for K600. See Data Source Format 
below.}

\item{velocdaily}{Data Source for daily mean flow velocity, for use in 
identifying daily priors or fixed values for K600. See Data Source Format 
below.}

\item{start_date}{NA or datetime, to be coerced with 
\code{as.POSIXct(start_date, tz="UTC")}, at which to start the data passed 
to the metab model}

\item{end_date}{NA or datetime, to be coerced with \code{as.POSIXct(end_date,
tz="UTC")}, at which to end the data passed to the metab model}

\item{tag}{character of form "1.0.2" that uniquely identifies this set of 
modeling runs.}

\item{strategy}{character, or vector of length sites, describing this set of 
modeling runs in concise English.}

\item{date}{POSIXct indicating the date of config construction. It is 
strongly recommended to use the default.}

\item{model}{character. the name of the metabolism model to construct}

\item{model_args}{character, in R language, specifying any arguments to pass 
to the model function}

\item{simplify}{logical. If there's only one data.frame to retrieve, remove 
it from the enclosing list?}

\item{verbose}{logical. give progress messages?}
}
\value{
a list of data.frames, or if simplify=TRUE and there's only one
  data.frame, a single data.frame of input data
}
\description{
Quick access to metabolism input data (\code{data} arg only)
}
\examples{
\dontrun{
md <- get_metab_data('nwis_08062500')
md <- get_metab_data('styx_004001')
login_sb()
md <- get_metab_data('styx_004001', doobs=choose_data_source(
  'doobs','styx_004001','sim',type='pred',src='160411 0.0.18'))
}
}

