% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{get site IDS from sciencebase}
\usage{
get_sites(session = NULL, limit = 10000)
}
\arguments{
\item{session}{arguments passed to sbtools \code{query_item_identifier} for authentication.
Not needed if parent of sites is public}

\item{limit}{numeric. Max number of sites to return. Default 10000}
}
\value{
a character vector of "site_root" titles (keys)
}
\description{
collects all "site_root" titles from sciencebase, returns as a character vector
}
\examples{
\donttest{
get_sites()
}
\dontrun{
# get sites if they are private
session <- authenticate_sb(username, password)
get_sites(session = session, limit = 10)
}
}

