% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{A helper to list_sites. Gets site IDS from sciencebase w/ simple data query}
\usage{
get_sites(with_dataset_name = NULL, with_ts_version = "rds",
  with_ts_archived = FALSE, with_ts_uploaded_after = "2015-01-01",
  limit = 10000)
}
\arguments{
\item{with_dataset_name}{limit sites to those with children matching the 
specified ts or other dataset name (e.g., "ts_doobs_nwis")}

\item{with_ts_version}{one or more of \code{c('rds','tsv')} to limit the 
dataset extension to anything in with_ts_version (if the dataset is a ts)}

\item{with_ts_archived}{one or more of \code{c(TRUE,FALSE)} to limit the list
to sites that have a ts that's archived, not archived, or either}

\item{with_ts_uploaded_after}{POSIXct, or convertible to POSIXct, giving date
after which a ts must have been uploaded to count}

\item{limit}{integer. the maximum number of items to return}
}
\value{
a character vector of "site_root" titles (keys)
}
\description{
Collects from sciencebase all site names ("site_root" titles, e.g., 
"nwis_02948375") or the names of sites containing a particular var_src 
dataset, and returns as a character vector of those site names
}
\examples{
\dontrun{
mda.streams:::get_sites()
mda.streams:::get_sites(limit = 10)
# get those sites that have water temperature in rds, non-archive form
mda.streams:::get_sites(with_dataset_name='ts_disch_nwis', with_ts_version='tsv')
mda.streams:::get_sites(with_dataset_name='ts_doobs_nwis', 
  with_ts_version=c('tsv','rds'), with_ts_archived=TRUE)
}
}
\keyword{internal}
