% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{A helper to list_sites. Gets site IDS from sciencebase w/ simple data query}
\usage{
get_sites(with_dataset_name = NULL, limit = 10000)
}
\arguments{
\item{with_dataset_name}{limit sites to those with children matching the
specified ts or other dataset name (e.g., "ts_doobs_nwis")}

\item{limit}{numeric. Max number of sites to return}
}
\value{
a character vector of "site_root" titles (keys)
}
\description{
Collects from sciencebase all site names ("site_root" titles, e.g., 
"nwis_02948375") or the names of sites containing a particular var_src 
dataset, and returns as a character vector of those site names
}
\examples{
\dontrun{
mda.streams:::get_sites()
mda.streams:::get_sites(limit = 10)
# get those sites that have water temperature
mda.streams:::get_sites(with_dataset_name = 'ts_wtr_nwis')
}
}
\keyword{internal}

