% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_var_codes.R
\name{get_var_codes}
\alias{get_var_codes}
\title{get project codes}
\usage{
get_var_codes(shortname, out = c("p_code", "units", "sm_name", "longname",
  "sources"), drop = TRUE)
}
\arguments{
\item{shortname}{a valid shortname for project (e.g., 'disch')}

\item{out}{character or character vector of name types to include in output.
If there is only one, the result will be a vector/value. One or more of
these options may be selected:
\code{c("shortname","p_code","sm_name","longname")} (where \code{shortname}
is also the rowname)}

\item{drop}{logical. In data.frame subsetting, should a single column be
dropped to a vector or a single row be dropped to a list?}
}
\value{
vector or data.frame of the requested variables and outnames.
  \code{shortname} is the name used in mda.streams. \code{p_code} is the
  corresponding code in the NWIS or NLDAS databases. \code{sm_name} is the
  name used in streamMetabolizer. \code{longtime} is the human-readable
  description.
}
\description{
a lookup for parameter codes and longnames based on the
  shortname convention used in mda.streams
}
\examples{
get_var_codes()
get_var_codes('wtr')
get_var_codes('disch', 'p_code')
get_var_codes('baro', c('longname','p_code'))
}

