% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_watershed_WFS.R
\name{get_watershed_WFS}
\alias{get_watershed_WFS}
\title{get WFS from site}
\usage{
get_watershed_WFS(site, session = NULL)
}
\arguments{
\item{site}{a site identifier}

\item{session}{arguments passed to sbtools \code{query_item_identifier} for authentication.
Not needed if parent of sites is public}
}
\value{
a character string for WFS url or NULL if watershed doesn't exist for site, or site doesn't exist.
}
\description{
get a WFS url from a site that has a 'watershed'
}
\examples{
site <- "nwis_04165500"
get_watershed_WFS(site)

\dontrun{
# will fail
get_watershed_WFS('fake_site'))
}
}
\seealso{
\code{\link{get_watershed_WMS}}
}

