% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_watershed_WMS.R
\name{get_watershed_WMS}
\alias{get_watershed_WMS}
\title{get WMS from site}
\usage{
get_watershed_WMS(site, ...)
}
\arguments{
\item{site}{a site identifier}

\item{...}{Additional parameters supplied to \code{\link[sbtools]{session_check_reauth}}}
}
\value{
a character string for WMS url or NULL if watershed doesn't exist for site, or site doesn't exist.
}
\description{
get a WMS url from a site that has a 'watershed'
}
\examples{
site <- "nwis_04165500"
get_watershed_WMS(site)

\dontrun{
# will fail
get_watershed_WMS('fake_site'))
}
}
\seealso{
\code{\link{get_watershed_WFS}}
}

