% Generated by roxygen2 (4.0.2): do not edit by hand
\name{init_nwis_sites}
\alias{init_nwis_sites}
\title{generate nwis site_ids for given p_codes}
\usage{
init_nwis_sites(p_codes)
}
\arguments{
\item{p_codes}{a character vector of NWIS p_codes}
}
\value{
a character vector of NWIS sites, appended with 'nwis_'
}
\description{
finds all NWIS sites that meet given data requirements
}
\examples{
gen_nwis_sites(p_codes)
}

