% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Get a vector of timeseries variable names}
\usage{
list_datasets(site, type = c("ts", "watershed"), ...)
}
\arguments{
\item{site}{a valid mda.streams site (see \link{get_sites}) or NULL for all
sites}

\item{type}{character. one or more dataset types to return}

\item{...}{additional arguments passed to \code{\link[sbtools]{session_check_reauth}},
  for example \code{username}}
}
\value{
an alphabetically sorted character vector of unique timeseries
  variable names for given sites
}
\description{
list_datasets() returns a list of all possible timeseries variables, while
list_datasets(site) returns the data available to a specific site.
}
\examples{
\dontrun{
list_datasets(site = 'nwis_01018035')
}
}

