% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Get a vector of timeseries dataset names}
\usage{
list_datasets(site_name, type = c("ts", "watershed"), ...)
}
\arguments{
\item{site_name}{a character vector of length one with a site name such
as those returned from make_site_name()}

\item{type}{character. one or more dataset types to return}

\item{...}{additional arguments passed to
  \code{\link[sbtools]{query_item_identifier}}, for example \code{limit}}
}
\value{
an alphabetically sorted character vector of unique timeseries
  variable names for given sites
}
\description{
list_datasets(site) returns the data available to a specific site. In
contrast, get_var_codes() returns a list of all possible variables and
get_var_codes(type="ts") returns all possible timeseries variables.
}
\examples{
\dontrun{
list_datasets(site_name = 'nwis_01021050')
}
}

