% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Get a vector of timeseries dataset names}
\usage{
list_datasets(site_name, data_type = c("ts", "watershed"),
  with_ts_version = "rds", with_ts_archived = FALSE,
  with_ts_uploaded_after = "2015-01-01", limit = 10000)
}
\arguments{
\item{site_name}{a character vector of length one with a site name such as 
those returned from make_site_name()}

\item{data_type}{character. one or more dataset types to return}

\item{with_ts_version}{one or more of \code{c('rds','tsv')} to limit the 
dataset extension to anything in with_ts_version (if the dataset is a ts)}

\item{with_ts_archived}{one or more of \code{c(TRUE,FALSE)} to limit the list
to sites that have a ts that's archived, not archived, or either}

\item{with_ts_uploaded_after}{POSIXct, or convertible to POSIXct, giving date
after which a ts must have been uploaded to count}

\item{limit}{integer. the maximum number of items to return}
}
\value{
an alphabetically sorted character vector of unique timeseries 
  variable names (in var_src format)for given sites
}
\description{
list_datasets(site) returns the data available to a specific site. In 
contrast, \code{get_var_src_codes(out="var_src")} returns a list of all
possible variables.
}
\examples{
\dontrun{
list_datasets(site_name = 'nwis_01021050')
}
}

