% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_metab_models.R
\name{list_metab_models}
\alias{list_metab_models}
\title{List the available metab_model objects}
\usage{
list_metab_models(order_by = c("date", "tag", "row", "site", "strategy",
  "title"))
}
\arguments{
\item{order_by}{character vector of aspects of the model names to sort on.
Options are the same as those in the \code{out} argument to
\code{\link{parse_metab_model_name}}}
}
\value{
a character vector of titles of the metab_model .RData files posted
  on SB
}
\description{
List the available metab_model objects
}

