% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_sites.R
\name{list_sites}
\alias{list_sites}
\title{Get a list of the sites on ScienceBase}
\usage{
list_sites(with_var_src = NULL, logic = c("all", "any"),
  with_ts_version = "rds", with_ts_archived = FALSE,
  with_ts_uploaded_after = "2015-01-01", limit = 10000)
}
\arguments{
\item{with_var_src}{character vector of data variables (i.e., 1+ of those 
listed in get_var_src_codes(out='var_src')), or a list of character vectors
where list names are 'any' and/or 'all' to define a combination of logical 
requirements}

\item{logic}{how to join the constraints in with_var_src, ...: is any of the 
listed parameters sufficient, or do you need all of them to be available 
for a site to qualify?}

\item{with_ts_version}{one or more of \code{c('rds','tsv')} to limit the 
dataset extension to anything in with_ts_version (if the dataset is a ts)}

\item{with_ts_archived}{one or more of \code{c(TRUE,FALSE)} to limit the list
to sites that have a ts that's archived, not archived, or either}

\item{with_ts_uploaded_after}{POSIXct, or convertible to POSIXct, giving date
after which a ts must have been uploaded to count}

\item{limit}{integer. the maximum number of items to return}
}
\value{
a character vector of site IDs
}
\description{
The with_var_src argument optionally limits the list to those sites that 
contain specific timeseries variables.
}
\examples{
\dontrun{
list_sites()
list_sites(with_var_src=c("wtr_nwis","doobs_nwis"), logic="any")
list_sites(with_var_src=c("wtr_nwis","doobs_nwis"), logic="any",
  with_ts_version=c('tsv'), with_ts_archived=TRUE, limit=45)
list_sites(list("wtr_nwis",any=c("doobs_nwis","doobs_simCopy")), logic="all")
}
}
\seealso{
\code{\link{locate_site}} \code{\link{list_tses}}
}

