% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_item.R
\name{locate_folder}
\alias{locate_folder}
\title{Find a high-level folder on ScienceBase}
\usage{
locate_folder(folder = c("project", "presentations", "proposals",
  "publications", "sites"), format = c("id", "url"), ...)
}
\arguments{
\item{folder}{the folder to locate}

\item{format}{character indicating whether the folder should be returned as
an ID or as a full URL}

\item{...}{other args passed to \code{\link[sbtools]{session_check_reauth}}}
}
\description{
Find a high-level folder on ScienceBase
}
\examples{
\dontrun{
locate_folder("publications", format="url")
testthat::expect_error(locate_folder("cvs", format="url"))
}
}

