% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_metab_run.R
\name{locate_metab_run}
\alias{locate_metab_run}
\title{Find a metab_run item on ScienceBase}
\usage{
locate_metab_run(title, format = c("id", "url"), by = c("tag", "dir",
  "either"), limit = 5000, browser = (format == "url"))
}
\arguments{
\item{title}{the title of the metabolism modeling run you want. this is the
date, tag, and run strategy separated by spaces.}

\item{format}{character indicating whether the folder should be returned as 
an ID or as a full URL}

\item{by}{character indicating how to search for the item: using tags ("tag",
the default and recommended option), by scanning the parent directory for 
the desired title ("dir"), or both in combination ("either")?}

\item{limit}{number of items to return, as in 
\code{\link[sbtools]{query_item_identifier}} or 
\code{\link[sbtools]{item_list_children}}}

\item{browser}{logical. Should the URL be opened in a browser?}
}
\description{
Find a metab_run item by title
}
\examples{
\dontrun{
locate_metab_run('150714 0.0.2 local_makefile_run')
}
}

