% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_site.R
\name{locate_site}
\alias{locate_site}
\title{Find a site folder on ScienceBase}
\usage{
locate_site(site_name, format = c("id", "url"), by = c("tag", "dir",
  "either"), limit = 5000, browser = (format == "url"))
}
\arguments{
\item{site_name}{the site ID, e.g. "nwis_02322688", whose folder you want}

\item{format}{character indicating whether the folder should be returned as
an ID or as a full URL}

\item{by}{character indicating how to search for the item: using tags ("tag",
the default and recommended option), by scanning the parent directory for
the desired title ("dir"), or both in combination ("either")?}

\item{limit}{number of items to return, as in
\code{\link[sbtools]{query_item_identifier}} or
\code{\link[sbtools]{item_list_children}}}

\item{browser}{logical. Should the URL be opened in a browser?}
}
\description{
Find a site folder on ScienceBase
}
\examples{
\dontrun{
locate_site("nwis_02322688", format="url")
locate_site(c("nwis_02322688", "nwis_03259813", "nwis_04024000"))
locate_site("nwis_notasite", format="url")
testthat::expect_error(locate_site("notasite", format="url"))
}
}

