% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_item.R
\name{locate_site}
\alias{locate_site}
\title{Find a site folder on ScienceBase}
\usage{
locate_site(siteid, format = c("id", "url"), ...)
}
\arguments{
\item{siteid}{the site ID, e.g. "nwis_02322688", whose folder you want}

\item{format}{character indicating whether the folder should be returned as
an ID or as a full URL}

\item{...}{other args passed to \code{\link[sbtools]{session_check_reauth}}}
}
\description{
Find a site folder on ScienceBase
}
\examples{
\dontrun{
locate_site("nwis_02322688", format="url")
locate_site(c("nwis_02322688", "nwis_03259813", "nwis_04024000"))
locate_site("nwis_notasite", format="url")
testthat::expect_error(locate_site("notasite", format="url"))
}
}

