% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/locate_item.R
\name{locate_ts}
\alias{locate_ts}
\title{Find a timeseries item on ScienceBase}
\usage{
locate_ts(variable = "doobs", siteid = "nwis_02322688", format = c("id",
  "url"), ...)
}
\arguments{
\item{variable}{the variable name, e.g., "doobs"}

\item{siteid}{the site ID, e.g. "nwis_02322688", whose folder you want}

\item{format}{character indicating whether the folder should be returned as
an ID or as a full URL}

\item{...}{other args passed to \code{\link[sbtools]{session_check_reauth}}}
}
\description{
Find a timeseries item on ScienceBase
}
\examples{
\dontrun{
locate_ts(c("doobs","wtr","disch"), "nwis_02322688")
locate_ts("doobs", "nwis_02322688", format="url")
locate_ts("doobs", "nwis_notasite", format="url")
testthat::expect_error(locate_ts("notavar", "nwis_notasite"))
}
}

