% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_sb.R
\name{login_sb}
\alias{login_sb}
\title{Log into ScienceBase with your myUSGS credentials}
\usage{
login_sb(username)
}
\arguments{
\item{username}{Your ScienceBase/myUSGS username, usually an email address}
}
\description{
BE CAREFUL NOT TO POST OR SHARE YOUR PASSWORD! Use the same username and 
password you set up via the ScienceBase interface or your government 
employer. This function is equivalent to 
\code{sbtools::\link[sbtools]{authenticate_sb}} but easier to type.
}

