% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_metab_model_path}
\alias{make_metab_model_path}
\title{Combine the model_name and folder into a file path for a metab_model}
\usage{
make_metab_model_path(model_name, folder, version = c("original", "modern"))
}
\arguments{
\item{model_name}{the name of the model as from make_metab_model_name()}

\item{folder}{the folder of the model file}
}
\description{
Combine the model_name and folder into a file path for a metab_model
}

