% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_ts_archive_path}
\alias{make_ts_archive_path}
\title{Create an archive file name for a file on ScienceBase}
\usage{
make_ts_archive_path(filename, creation_date)
}
\arguments{
\item{filename}{of the file that is to be archived}

\item{creation_date}{the creation date of the ScienceBase file that will be 
archived (as.Date)}
}
\value{
a file name to be used as an archive
}
\description{
Create an archive file name for a file on ScienceBase
}

