% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_ts_name}
\alias{make_ts_name}
\title{Translate a timeseries name from mda.streams to ScienceBase}
\usage{
make_ts_name(variable)
}
\arguments{
\item{variable}{a timeseries name[s] in mda.streams lingo (e.g., \code{wtr})}
}
\value{
timseries name[s] in ScienceBase lingo (e.g., "\code{ts_wtr})
}
\description{
Translate a timeseries name from mda.streams to ScienceBase
}

