% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_ts_path}
\alias{make_ts_path}
\title{Create a standardized file name for specific file contents}
\usage{
make_ts_path(site_name, ts_name, folder, version = c("tsv", "rds", "RData"))
}
\arguments{
\item{site_name}{the full site name, e.g., 'nwis_06893820'}

\item{ts_name}{the full ts name, e.g., 'ts_doobs_nwis'}

\item{folder}{the folder to write the file in, or missing}

\item{version}{character string indicating the file format}
}
\value{
a full file path
}
\description{
Create a standardized file name for specific file contents
}

