% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{make_var_src}
\alias{make_var_src}
\title{Translate a variable and data source into a var_src}
\usage{
make_var_src(var, src)
}
\arguments{
\item{var}{a variable shortname from
\code{unique(get_var_src_codes(out='var'))}, e.g., "doobs"}

\item{src}{a data source from the sources available for the given var; e.g., 
see \code{get_var_src_codes(var=='doobs', out='src')}}
}
\description{
Translate a variable and data source into a var_src
}

