% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_ts_name}
\alias{parse_ts_name}
\title{Translate a timeseries name from ScienceBase to mda.streams}
\usage{
parse_ts_name(ts_name, out = "var_src", use_names = length(out) > 1)
}
\arguments{
\item{ts_name}{timeseries name[s] in ScienceBase lingo (e.g.,
\code{ts_wtr_nwis})}

\item{out}{character, length 1 or 2 selected from \code{c("var_src",
"variable", "src")} indicating whether you want to get back the
joined variable & source, the variable, the src, or some
combination.}

\item{use_names}{logical. Should the return vector or data.frame be named
according to the input values?}
}
\value{
timeseries name[s]in mda.streams lingo (e.g., \code{wtr})
}
\description{
Translate a timeseries name from ScienceBase to mda.streams
}

