% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_ts_name}
\alias{parse_ts_name}
\title{Translate a timeseries name from ScienceBase to mda.streams}
\usage{
parse_ts_name(ts_name, out = "var_src", use_names = length(ts_name) > 1)
}
\arguments{
\item{ts_name}{timeseries name[s] in ScienceBase lingo (e.g., 
\code{ts_wtr_nwis})}

\item{out}{character, length 1 or 2 selected from \code{c("var_src", 
"var", "src")} indicating whether you want to get back the 
joined variable & source, the variable, the src, or some 
combination.}

\item{use_names}{logical. Should the return vector or data.frame be named
according to the input values?}
}
\value{
timeseries name[s]in mda.streams lingo (e.g., \code{wtr})
}
\description{
Translate a timeseries name from ScienceBase to mda.streams
}
\examples{
mda.streams:::parse_ts_name("ts_doobs_nwis")
mda.streams:::parse_ts_name(c("ts_doobs_nwis", "ts_stage_nwis"))
mda.streams:::parse_ts_name(ts_name="ts_doobs_nwis", out=c("var_src","src"))
mda.streams:::parse_ts_name(c("ts_doobs_nwis", "ts_stage_nwis"), c("src","var"))
}

