% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mda_helpers.R
\name{parse_ts_path}
\alias{parse_ts_path}
\title{Split file path into contents}
\usage{
parse_ts_path(file_path, out = c("dir_name", "file_name", "site_name",
  "ts_name", "var_src", "var", "src", "database", "sitenum"),
  use_names = length(file_path) > 1)
}
\arguments{
\item{file_path}{a valid file path for ts file}

\item{out}{a character for desired output ('dir_name','file_name',
'site_name', 'ts_name', or any of the out names from parse_ts_name or
parse_site_name)}

\item{use_names}{logical. Should the return vector be named according to the
input values?}
}
\value{
a character
}
\description{
Split file path into contents
}

