% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_indy_site.R
\name{post_indy_site}
\alias{post_indy_site}
\title{Create a indy site and post a list of timeseries files to it}
\usage{
post_indy_site(file_list, on_site_exists = c("clear", "skip", "stop",
  "replace"), summarize = FALSE)
}
\arguments{
\item{file_list}{a list of files as given by stage_indy_site. The site name 
will be determined from the first item (a metadata vector) in this list.}

\item{on_site_exists}{see \code{on_exists} argument to 
\code{\link{post_site}}}

\item{summarize}{logical. return a summary of the site's tses?}
}
\description{
This function is easiest to use when applied to the output of 
\code{\link{stage_indy_site}}.
}
\examples{
\dontrun{
staged <- stage_indy_site(info="help file example for post_indy_site")
login_sb()
post_indy_site(staged)
}
}

