% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_site.R
\name{post_site}
\alias{post_site}
\title{Create new site in SB}
\usage{
post_site(sites, on_exists = c("stop", "skip", "clear", "replace"),
  verbose = TRUE)
}
\arguments{
\item{sites}{one or more local site IDs (e.g. nwis_09238475)}

\item{on_exists}{character. what should be done when an item already exists?}

\item{verbose}{logical. Should status messages be given?}
}
\value{
an item list
}
\description{
Create new site in SB
}
\examples{
\dontrun{
sbtools::authenticate_sb()
set_scheme("mda_streams_dev")

sites <- c("nwis_00000000", "nwis_00000001", "nwis_00000002")
is.null(post_site()) # returns NULL (empty list) if sites is missing or NULL
(post_site(sites)) # adds sites, returns item IDs
(post_site(sites, on_exists="skip")) # leaves existing items untouched, returns item IDs
(post_site(sites, on_exists="clear")) # empties existing sites of child items, returns item IDs
mda.streams:::delete_site(sites)

set_scheme("mda_streams")
}
}
\author{
Corinna Gries
}

