% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_ts.R
\name{post_ts}
\alias{post_ts}
\title{Post a new timeseries file to SB}
\usage{
post_ts(files, ...)
}
\arguments{
\item{files}{a string vector of file paths for POSTing}

\item{...}{Session object from \link{authenticate_sb}}
}
\description{

}
\examples{
\dontrun{
files <- stage_nldas_ts(sites = c("nwis_06893820","nwis_01484680"), variable = "baro",
                times = c('2014-01-01 00:00','2014-01-01 05:00'))
post_ts(files, session = sbtools::authenticate_sb())
}
}
\author{
Luke Winslow, Corinna Gries, Jordan S Read
}

