% Generated by roxygen2 (4.0.2): do not edit by hand
\name{post_ts}
\alias{post_ts}
\title{Post a new timeseries file to SB}
\usage{
post_ts(site, data, variable, session)
}
\arguments{
\item{site}{Unique site ID}

\item{data}{A data.frame containing the timeseries}

\item{variable}{Timeseries identifiying string variable eg [stage, doobs, wtr, etc]}

\item{session}{Session object from \link{authenticate_sb}}
}
\description{

}
\author{
Luke Winslow, Corinna Gries
}

