% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_ts.R
\name{post_ts}
\alias{post_ts}
\title{Post a new timeseries file to SB}
\usage{
post_ts(files, on_exists = c("stop", "skip", "replace", "merge"),
  verbose = TRUE)
}
\arguments{
\item{files}{a string vector of file paths for POSTing}

\item{on_exists}{character specifying an action when a file to post already
exists on ScienceBase}

\item{verbose}{logical. Should status messages be given?}
}
\description{
Post a staged file from the local computer to ScienceBase
}
\examples{
\dontrun{
login_sb()
set_scheme("mda_streams_dev")

sites <- c("nwis_05406479", "nwis_05435950", "nwis_04087119")
post_site(sites, on_exists="clear")
files <- stage_nwis_ts(sites = sites, var = "doobs",
  times = c('2014-01-01 00:00','2014-01-01 05:00'))
post_ts(files)
locate_ts("doobs_nwis", sites) # find the newly posted data online
post_ts(files, on_exists="skip")
post_ts(files, on_exists="replace")
post_site(sites, on_exists="clear")

set_scheme("mda_streams_")
}
}
\author{
Luke Winslow, Corinna Gries, Jordan S Read
}

