% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/post_ts.R
\name{post_ts}
\alias{post_ts}
\title{Post a new timeseries file to SB}
\usage{
post_ts(site, data, variable_name, session)
}
\arguments{
\item{site}{Unique site ID}

\item{data}{A data.frame containing the timeseries}

\item{variable_name}{Timeseries identifiying string variable_name eg [stage, doobs, wtr, etc]}

\item{session}{Session object from \link{authenticate_sb}}
}
\description{

}
\author{
Luke Winslow, Corinna Gries
}

