% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_watershed.R
\name{post_watershed}
\alias{post_watershed}
\title{Post a new watershed shape to a site on SB}
\usage{
post_watershed(site, files)
}
\arguments{
\item{site}{The unique site identifie?r}

\item{files}{String vector of files that make up the shapefile}
}
\value{
Returns SB id of newly created watershed item
}
\description{
Posts a file from the local computer to ScienceBase.
}
\author{
Luke Winslow
}

