% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/post_watershed.R
\name{post_watershed}
\alias{post_watershed}
\title{Post a new watershed shape to a site on SB}
\usage{
post_watershed(site, files, session)
}
\arguments{
\item{site}{The unique site identifie?r}

\item{files}{String vector of files that make up the shapefile}

\item{session}{A SB session from \link{authenticate_sb}}
}
\value{
Returns SB id of newly created watershed item
}
\description{
Posts a file from the local computer to ScienceBase.
}
\author{
Luke Winslow
}

