% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ts.R
\name{read_ts}
\alias{read_ts}
\title{Read timeseries data into a data.frame}
\usage{
read_ts(file, on_invalid = c("stop", "warn"))
}
\arguments{
\item{file}{a valid ts file path}

\item{on_invalid}{character in \code{c("stop","warn")} indicating how to
handle invalid timeseries}
}
\value{
a timeseries unitted data.frame
}
\description{
Reads a ts file (with format specific to mda.streams) into a data.frame
}
\examples{
\dontrun{
file <- download_ts(var_src = "baro_nldas", site_name = "nwis_06893820")
baro_pressure <- read_ts(file)
}
}
\author{
Jordan S. Read, Alison Appling
}
\seealso{
\code{\link{download_ts}}
}

