% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_meta.R, R/read_ts.R
\name{read_ts}
\alias{read_ts}
\title{Read metadata file for mda.streams into data.frame}
\usage{
read_ts(file)

read_ts(file)
}
\arguments{
\item{file}{a valid ts file path}

\item{file}{a valid ts file path}
}
\value{
a metadata data.frame

a timeseries unitted data.frame
}
\description{
Read metadata file for mda.streams into data.frame

read ts format (timeseries) for mda.streams into data.frame
}
\examples{
\dontrun{
file <- download_meta(type='basic')
meta_basic <- read_meta(file)
}
\dontrun{
file <- download_ts(var_src = "baro_nldas", site_name = "nwis_06893820")
baro_pressure <- read_ts(file)
}
}
\author{
Alison P. Appling

Jordan S. Read
}
\seealso{
\code{\link{download_meta}}

\code{\link{download_ts}}
}

