% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_ts.R
\name{read_ts}
\alias{read_ts}
\title{read ts format (timeseries) for mda.streams into data.frame}
\usage{
read_ts(file_handle)
}
\arguments{
\item{file_handle}{a valid ts file handle}
}
\value{
a timeseries data.frame
}
\description{
read ts format (timeseries) for mda.streams into data.frame
}
\examples{
\dontrun{
file_handle <- download_ts(site = 'nwis_01018035', variable = 'doobs')
dissolved_oxygen <- read_ts(file_handle)
}
}

