% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_calc_ts.R
\name{stage_calc_ts}
\alias{stage_calc_ts}
\title{stage calculated or modeled data into a time series file}
\usage{
stage_calc_ts(sites, variable, folder = tempdir(), verbose = FALSE, ...)
}
\arguments{
\item{sites}{a character vector of valid NWIS site IDs}

\item{variable}{short name of variable as in \code{\link{get_var_codes(out='shortname')}}}

\item{folder}{a folder to place the file outputs in (defaults to temp directory)}

\item{verbose}{provide verbose output (currently not implemented)}

\item{...}{additional arguments passed to \code{\link[geoknife]{geoknife}} and \code{\link[unitted]{write_unitted}}}
}
\value{
a file handle for time series file created
}
\description{
accept calculated/modeled data as a data.frame and return a file handle
}
\examples{
\dontrun{
files <- stage_calc_ts(sites = c("nwis_06893820","nwis_01484680"), variable = "baro",
                 times = c('2014-01-01 00:00','2014-01-01 05:00'))
}
}

