% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_meta_basic.R
\name{stage_meta_basic_indy}
\alias{stage_meta_basic_indy}
\title{Get data for Indy (custom/independently collected data) sites}
\usage{
stage_meta_basic_indy(sites_meta, empty_meta, verbose = FALSE)
}
\arguments{
\item{sites_meta}{}

\item{verbose}{logical. give status messages?}
}
\description{
Helper to stage_meta_basic. Copies data from meta_indy into meta_basic where
this is available.
}

