% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_meta_styx.R
\name{stage_meta_styx}
\alias{stage_meta_styx}
\title{Create/modify the meta_styx file}
\usage{
stage_meta_styx(rows, on_exists = c("stop", "replace"), folder = tempdir(),
  verbose = FALSE)
}
\arguments{
\item{rows}{a data frame with one or more rows, having the same columns as 
the existing meta_ file (if there is one). These rows will be inserted 
according to their site_name values.}

\item{on_exists}{character indicating how to handle new rows whose site_name 
is already present in the current meta_ file If 'stop', an error will be 
thrown if even one value is duplicated. If 'replace', a message will be 
given and existing rows with site_name in rows$site_name will be replaced
by the new rows.}

\item{folder}{where to store the staged file}

\item{verbose}{logical. give status messages?}
}
\description{
Create/modify the meta_styx file
}

