% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_metab_config.R
\name{stage_metab_config}
\alias{stage_metab_config}
\title{Write a metabolism modeling configuration file}
\usage{
stage_metab_config(tag, strategy, date = Sys.time(), model = "metab_mle",
  model_args = "list()", site = list_sites(c("doobs_nwis", "disch_nwis",
  "wtr_nwis")), sitetime = choose_data_source("sitetime", site),
  doobs = choose_data_source("doobs", site),
  dosat = choose_data_source("dosat", site),
  depth = choose_data_source("depth", site), wtr = choose_data_source("wtr",
  site), par = choose_data_source("par", site),
  sitedate = choose_data_source("sitedate", site, logic = "unused var"),
  K600 = choose_data_source("K600", site, logic = "unused var"),
  dischdaily = choose_data_source("dischdaily", site, logic = "unused var"),
  velocdaily = choose_data_source("velocdaily", site, logic = "unused var"),
  omit_incomplete = TRUE, filename = "./condor_config.tsv")
}
\arguments{
\item{tag}{character of form "1.0.2" that uniquely identifies this set of
modeling runs.}

\item{strategy}{character, or vector of length sites, describing this set of
modeling runs in concise English.}

\item{date}{POSIXct indicating the date of config construction. It is
strongly recommended to use the default.}

\item{model}{character. the name of the metabolism model to construct}

\item{model_args}{character, in R language, specifying any arguments to pass
to the model function}

\item{site}{site names}

\item{sitetime}{2-column data.frame with names "type" and "src" describing
where apparent solar time data should come from}

\item{doobs}{2-column data.frame with names "type" and "src" describing where
DO data should come from}

\item{dosat}{2-column data.frame with names "type" and "src" describing where
DO saturation data should come from}

\item{depth}{2-column data.frame with names "type" and "src" describing where
stream depth data should come from}

\item{wtr}{2-column data.frame with names "type" and "src" describing where
water temperature data should come from}

\item{par}{2-column data.frame with names "type" and "src" describing where
light (photosynthetically available radiation, PAR) data should come from}

\item{sitedate}{see mm_data}

\item{K600}{see mm_data}

\item{dischdaily}{see mm_data}

\item{velocdaily}{see mm_data}

\item{omit_incomplete}{logical. If one or more datasets required for the
specified config row is unavailable, should that row be omitted?}

\item{filename}{character or NULL. If NULL, the function returns a
data.frame, otherwise it writes that data.frame to the file specified by
filename.}
}
\value{
file name of the config file
}
\description{
Write a table (tsv) of configuration information for individual metabolism
modeling jobs (one row/job per site-strategy combination). This tsv should
reflect the full information needed to re-run a set of jobs. The jobs will
probably, but not necessarily, be run on a Condor cluster.
}
\examples{
\dontrun{
login_sb(); site="nwis_01646000"; stage_metab_config(tag="0.0.1", strategy="try stage_metab_config",
  model="metab_Kvpred", site=site, filename=NULL,
  sitetime=choose_data_source("sitetime", site, logic="manual", src="calcLon", type="ts"),
  doobs=choose_data_source("doobs", site, logic="unused var"),
  dosat=choose_data_source("dosat", site, logic="unused var"),
  depth=choose_data_source("depth", site, logic="unused var"),
  wtr=choose_data_source("wtr", site, logic="unused var"),
  par=choose_data_source("par", site, logic="unused var"),
  K600=choose_data_source("K600", site, logic="nighttime", src="0.0.6", type="pred"),
  dischdaily=choose_data_source("dischdaily", site, logic="manual", src="calcDMean", type="ts"),
  velocdaily=choose_data_source("velocdaily", site, logic="manual", src="calcDMean", type="ts"),
  omit_incomplete=FALSE)
stage_metab_config(tag="0.0.1", strategy="try stage_metab_config",
  site="nwis_01646000", filename=NULL)
stage_metab_config(tag="0.0.1", strategy="test write_metab_config",
  site=list_sites()[24:33], filename=NULL,
  omit_incomplete=FALSE)
stage_metab_config(tag="0.0.1", strategy="test write_metab_config",
  site=list_sites()[24:33], filename=NULL)
}
}

