% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stage_nwis_sitelist.R
\name{stage_nwis_sitelist}
\alias{stage_nwis_sitelist}
\title{generate nwis site_ids for given p_codes}
\usage{
stage_nwis_sitelist(p_codes, state_codes, folder = NULL, verbose = TRUE)
}
\arguments{
\item{p_codes}{a character vector of NWIS p_codes}

\item{state_codes}{character vector of state codes, or "all" for all data.}

\item{folder}{directory path, or NULL, indicating where to save the file or
(NULL) to return it as a character vector}

\item{verbose}{logical. Should status messages be given?}
}
\value{
a character vector of NWIS sites, appended with 'nwis_'
}
\description{
finds all NWIS sites that meet given data requirements
}
\examples{
\dontrun{
stage_nwis_sitelist(p_codes=c('00095','00060','00010','00300'),
  state_codes=c("wi","Michigan"))
sites <- stage_nwis_sitelist(p_codes=get_var_codes("doobs","p_code"),
  state_codes=c("all"), verbose=TRUE)
sites_file <- stage_nwis_sitelist(p_codes=get_var_codes("doobs","p_code"),
  state_codes=c("wi"), folder=tempdir()); readLines(sites_file)
}
}

