% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_nwis_sitelist.R
\name{stage_nwis_sitelist}
\alias{stage_nwis_sitelist}
\title{generate nwis site_ids for given p_codes}
\usage{
stage_nwis_sitelist(vars, min.obs = 100, site.types = "ST", HUCs = 1:21,
  folder = NULL, verbose = TRUE, p_codes)
}
\arguments{
\item{vars}{a character vector of mda.streams var codes, as in 
\code{unique(get_var_src_codes(out='var'))}}

\item{min.obs}{minimum number of observations of the vars that a site must 
have to be accepted}

\item{site.types}{character vector of acceptable NWIS site types as in 
https://maps.waterdata.usgs.gov/mapper/help/sitetype.html}

\item{HUCs}{character or numeric vector of HUC codes, e.g., 1:3 or '04'}

\item{folder}{directory path, or NULL, indicating where to save the file or 
(NULL) to return it as a character vector}

\item{verbose}{logical. Should status messages be given?}

\item{p_codes}{optional - allows you to specify p_codes that aren't in 
\code{get_var_src_codes(!is.na(p_code), out=c("var","src","p_code"))} for 
exploratory purposes}
}
\value{
a character vector of NWIS sites, appended with 'nwis_'
}
\description{
finds all NWIS sites that meet given data requirements
}
\examples{
\dontrun{
stage_nwis_sitelist(vars=c('doobs','wtr','disch','stage'), 
  state_codes=c("wi","Michigan"))
sites <- stage_nwis_sitelist(
  p_codes=get_var_src_codes(var=="doobs",!is.na(p_code),out="p_code"), 
  state_codes=c("all"), verbose=TRUE)
sites_file <- stage_nwis_sitelist(
  p_codes=get_var_src_codes(var=="doobs",!is.na(p_code),out="p_code"), 
  state_codes=c("wi"), folder=tempdir()); readLines(sites_file)
}
}
