% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stage_nwis_ts.R
\name{stage_nwis_ts}
\alias{stage_nwis_ts}
\title{stage nwis data into a file}
\usage{
stage_nwis_ts(sites, var, times, folder = tempdir(), verbose = FALSE)
}
\arguments{
\item{sites}{a character vector of valid NWIS site IDs}

\item{var}{short name of variable as in
\code{unique(get_var_src_codes(out="var"))}}

\item{times}{a length 2 vector of text dates in YYYY-MM-DD format}

\item{folder}{a folder to place the file outputs in (defaults to temp 
directory)}

\item{verbose}{provide verbose output (currently not implemented)}
}
\value{
a character vector of file handles
}
\description{
get data from nwis and return created file handle
}
\examples{

\dontrun{

files <- stage_nwis_ts(sites = c("nwis_06893820","nwis_01484680"), 
  var = "doobs", times = c('2014-01-01','2014-01-02'), verbose=TRUE)
head(read_ts(files[1]))

# par is unavailable for all sites, so returns NULL
stage_nwis_ts(sites = list_sites(), var = "par",
  times = c('2014-01-01', '2014-01-03'), verbose=TRUE) 
}
}

