% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verify_ts.R
\name{verify_ts}
\alias{verify_ts}
\title{verify timeseries data.frame format and contents}
\usage{
verify_ts(data, var, checks = c("ncol", "unitted", "tz", "units", "names"),
  on_fail = warning)
}
\arguments{
\item{data}{a \code{\link[unitted]{unitted}} data.frame with date and value}

\item{checks}{tests to run on the data ('cols', 'tz', 'units', 'variable')}

\item{on_fail}{function to call when a test fails. good options are warning
and stop. if the function does not halt evaluation, FALSE is returned after
the function is first called.}

\item{var_src}{optional name for timeseries. If supplied, will check variable
column name}
}
\value{
TRUE if is valid, FALSE if not
}
\description{
Checks to see a given data.frame meets project criteria
}
\examples{
\dontrun{
files <- stage_nwis_ts(sites = c("nwis_06893820","nwis_01484680"), var = "doobs",
                 times = c('2014-01-01','2014-02-01'))
files <- download_ts("doobs_nwis", "nwis_01073389", on_local_exists="skip")
ts <- read_ts(files[1])
verify_ts(ts, var='doobs', checks=c('ncol','units','names')) # TRUE
verify_ts(ts, var='wtr', checks=c('ncol','units','names')) # warning + FALSE
}
}
\keyword{internal}

