% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts}
\alias{write_ts}
\title{Write timeseries data into a compressed file}
\usage{
write_ts(data, site, var, src, folder, version = c("tsv", "rds"))
}
\arguments{
\item{data}{unitted data.frame}

\item{site}{the full site name, e.g., 'nwis_06893820'}

\item{var}{the variable name of the output, e.g., 'doobs'}

\item{src}{the source of the data, e.g., 'nwis' or 'nldas'}

\item{folder}{the folder to write the file in}

\item{version}{the output version for the file (tsv, rds)}
}
\description{
Writes a ts file (with format specific to mda.streams) a data.frame into a
compressed file suitable for posting to ScienceBase
}
\keyword{internal}

