% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write_ts.R
\name{write_ts}
\alias{write_ts}
\title{write timeseries format compressed file}
\usage{
write_ts(data, site, variable, folder)
}
\arguments{
\item{data}{unitted data.frame}

\item{site}{the full site name (e.g., 'nwis_06893820')}

\item{variable}{the variable name of the output}

\item{folder}{the folder to write the file in}
}
\description{
write timeseries format compressed file
}
\keyword{internal}

